applications: asset_tracker_v2: Remove PSM handling from the application
Remove PSM handling from the application and depend on the
Link Controller library to request PSM through the
`CONFIG_LTE_PSM_REQ` option.

This is to prevent conflict when having PSM be requested in
multiple places at the same time and to have a single point of handling.

Fixes NCSDK-21774

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>