applications: nrf_desktop: Allow to suspend device when USB powered
Change allows to suspend device when in USB powered state. When an USB
cable is connected, the USB switches between powered and suspended state
a few times. Enforcing active power manager state on USB powered
triggers a sequence of multiple power downs and wake ups when an USB
cable is connected.

Jira: NCSDK-21875

Signed-off-by: Marek Pieta <Marek.Pieta@nordicsemi.no>
Signed-off-by: Divya Pillai <divya.pillai@nordicsemi.no>