CMake: doc: Add missing required env var ZEPHYR_NRF_MODULE_DIR
The nrf_security nrfxlib Kconfig sources willl source nrf Kconfigs and
therefore need to know where ZEPHYR_NRF_MODULE_DIR is. It's not clear
how other Kconfig env vars are added and why this needed to be added
as a special case here. There could be a better way to do this, but
the author does not know how.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>