nrf_security: Improve nrf_security_debug for depends configuration
The macro kconfig_check_and_set_base_to_val_depends has poor debug
logging information. It only logs when a configuration is "unset", i.e
when it did not enable a symbol because one of the dependencies were not
met. However it does not log when it sets a symbol to true.
This causes incomplete logging of which symbols are enabled.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>