lib: identity_key: Remove k_panic and return error
The identity key library until now panicked with k_panic
in case of error in most of the cases. It is a preferred
practice to return error codes and let the application
decide what to do with the error. This updates the functions
from void to int and adds some extra error codes to cover
all the cases.

Ref: NCSDK-21839

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>