lib: nrf_modem: rename `lte_connectivity` to `lte_net_if`
`lte_connectivity` is not a good name for this component.
It is possible to have full LTE connectivity without it.

Rather, what it does is provide support for Zephyr net_if API,
allowing the application to control the nRF91 as a Zephyr
network interface.

Rename to componenent `lte_net_if` and prefix Kconfig with `NRF_MODEM_LIB`
to indicate that it belongs to the nrf_modem library NCS integration
layer.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>