lib: nrf_modem: deprecate `NRF_MODEM_LIB_NET_IF_AUTO_START`
NCS shall not support the auto-initialization of the nRF91 modem during
system init (`SYS_INIT`). The initialization of the modem is a blocking
operation that can take several minutes, during which the application
has not even reached main() and is unable to do anything.

Further, the initialization result must be known to the application,
but `SYS_INIT` does not return anything to the application.

This is simply not acceptable in real products, and we shall not
provide such an option in NCS libraries. If the user wants, they can
easily add that themselves.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>