samples: psa_tls: Enable the default set of PSA want algs
Explicitly enable the default set of PSA want algorithms in the
psa_tls sample.
The set of algorithms is not neccesarily the required set of algorithms
but instead just what was already enabled as a simplifiction in order
to not cause any configuration changes in the sample.
Remove PSA want configuration from other kconfig fragments.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>