manifest: Update TF-M with max hash size fix
Update TF-M to include maximum hash size define fix.
Because the PSA_WANT configurations are not available in the TF-M
PSA configurations we always set the maximum hash size to 32 bytes.
When using SHA sizes over 256 bits this is not enough.

This update will default the maximum hash size define to 64 bytes when
the PSA_WANt configurations are not available, or use the PSA_WANT
configurations when they are available.

NCSDK-22273

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>