lib: nrf_modem: release fd lock in sendto()
Let sendto() release the fd lock to be able to execute other
operations in the meantime, for example, recv().

If sendto() doesn't release its lock, it can't be preempted by other
threads, including threads that recv(). This can create a deadlock
situation in case a lot of network data is received during a sendto()
call, filling up the shared memory and preventing the modem from
sending a response to the sendto() request, which then never returns,
preventing the recv() thread from being scheduled.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>