nrf_security: Adapt PSA_NEED configurations
The Oberon core uses a configuration scheme where the user
chooses the PSA_WANT and the PSA_USE options and the core itself
derives the PSA_NEED configurations which are used internally.

This change adapts this new configuration scheme in nrf_security.

With this change the user still have the option to explicitely
enable/disable drivers with the previous options:
PSA_CRYPTO_DRIVER_OBERON
PSA_CRYPTO_DRIVER_CC3XX

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>