Applications: nrf5340_audio: Add display module
Adds the nrf5340_audio_dk_display module for the audio application.
This module uses the LVGL library to present information from the
application to the user using an Adafruit touch display. A shield
definition is added so that the Adafruit display shield can be used
with the nrf5340 Audio DK.

Signed-off-by: Philip Fidjeland <philipfidjeland@gmail.com>
Signed-off-by: Andreas Vibeto <andreas.vibeto@nordicsemi.no>