Bluetooth: Mesh: Check sensor_value array size
By the sensor API design, `CONFIG_BT_MESH_SENSOR_CHANNELS_MAX` shall be
equal to `channel_count` of a sensor with the most number of channels.
If the configuration is incorrect, it may not be noticed immediately and
out of bound read or write can happen.

Add assertion to check that `CONFIG_BT_MESH_SENSOR_CHANNELS_MAX` is less
than or equal to the number of sensor channels being encoded or decoded.

The assertion is required for encoding as well because server allocates
`CONFIG_BT_MESH_SENSOR_CHANNELS_MAX` array to avoid usage of VLA based
on `channel_count` value of a specific sensor.

NCSDK-22553

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>