Bluetooth: Mesh: Drop sensor property on failure
When Sensor Server tries to encode sensor value, the value provided by
a sensor may be outside of the format range (e.g. format expects values
from 0 to 65535, but sensor returned 65536). sensor_status_encode()
could have already encoded Marshalled Property ID, but then return
error if the value is outside the format range. The specification
doesn't clearly define what to do in this case. But we can't return
property with zero length as it may be treated as not-supported property
id (from MshMDLv1.0.1, section 4.3.1.2.4, last paragraph):

or the Length shall represent the value of zero and the Raw Value
field shall contain only the Property ID if the requested device
property is not recognized by the Sensor Server.

This commit restores buf state after the failed encoding so the
corrupted value is not included in the message and the property is
ignored. The response is not send at all if none of sensor values were
encoded successfully.

NCSDK-21861 (sensor implementation fix).

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>