cmake: partition_manager: Generate images for S0/S1 to external flash
Implemented a fix in MCUBoot for NSIB + Multi-image for nRF53.

This enables partition manager and the build system to generate images
for S1 and S0 with the address of external QSPI flash. If the QSPI node
with XiP is defined in the DTS.

This is the same code used to move the application update into the
secondary slot in external flash.

Ref. NCSDK-19223

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>