tests: tfm: Remove builtin crypto configuration
The builtin crypto implementation, which is the software
implementation of PSA crypto APIs from mbedTLS is not
supported anymore. Only the Oberon PSA API implementation
exists right now as a software implementation, and there
are also the crypto accelerator implementation.

Remove the unused builtin PSA implementation configuration
in the tests.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>