tfm: Fix undefined reference to 'end' when disabling secure output
Fix undefined reference to 'end' when disabling secure output from the
TF-M image.

When secure output is disabled the printf function is not overriden
in TF-M and the libc variant has references to heap functions.
Make sure that the printf function is overriden for this case also and
drop the output message.

NCSDK-22818

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>