hostap: Implement offloaded Wi-Fi management ops
These ops are not implemented by the WPA supplicant and too much of a
hassle to implement a pass through all the way to the driver, so, we use
the offloaded APIs registered by the driver and make direct calls.

Though offloaded APIs are are registered as long as the interface is
managed by the WPA supplicant, Wi-Fi management will not invoke them
directly. But if WPA supplicant is disabled (Scan only mode), then Wi-Fi
management invokes them directly.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>