cmake: partition_manager: Replace import_kconfig
Following the changes proposed in:

https://github.com/zephyrproject-rtos/zephyr/pull/59915

the `import_kconfig()` function in upstream Zephyr has been updated to
become more aligned with Kconfiglib, which makes it less of a generic
"key=value" parser. This could provoke subtle bugs in the CMake bits of
partition manager, where the function is being used to parse `pm.config`
files, which are not actual Kconfig fragments.

Add a replacement parser for `pm.config` files.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>