boards: arm: thingy91: New static partition configuration options
Move static partition logic for thingy:91 out to separate Kconfig
options:

CONFIG_THINGY91_STATIC_PARTITIONS_FACTORY for the original factory
partition layout. This is the default.

CONFIG_THINGY91_STATIC_PARTITIONS_SECURE_BOOT for a partition layout
similar to the factory layout, but it has the immutable bootloader and
partitions for uptadable bootloader. This is still experimental.

CONFIG_THINGY91_STATIC_PARTITIONS_LWM2M_CARRIER for a partition layout
that has the partitions needed to use the lwm2m carrier library.

CIA-604

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>