lib: dk_buttons_and_leds: Add no interrupts mode for buttons
Add no interrupts mode for buttons (CONFIG_DK_LIBRARY_BUTTON_NO_ISR).
This option avoids relying on GPIO interrupts at all and unconditionally
periodically scan all of the buttons.

It can be used on the nRF54L15 PDK revision 0.2.X, where buttons 3 and
4 are connected to GPIO PORT2 which does not support GPIOTE.
Because of that, the PDK cannot rely on GPIO interrupts for the
mentioned buttons.

Jira: NCSDK-23210

Signed-off-by: Mateusz Kapala <mateusz.kapala@nordicsemi.no>