samples: bluetooth: Do not rely on buffer being freed when filled
This change updates the sample so that it doesn't rely on RX UART
buffer being freed only when fully filled with data. It might happen
that the buffer is freed by the driver even tho it's not fully filled.
In that case the data should not be send to NUS unless '\r' or '\n'
character was send.

This change adds additional buffering before sending to NUS.

Signed-off-by: Emil Obalski <Emil.Obalski@nordicsemi.no>