lib: lte_link_control: Handle NCELLMEAS response without request
When NCELLMEASSTOP is issued, lte_lc gives the semaphore indicating
that NCELLMEAS is not ongoing anymore.
However, NCELLMEAS notification is sent after NCELLMEASSTOP and it
depends on which thread the lte_lc_neighbor_cell_measurement_cancel()
is done whether NCELLMEAS notification is coming before or after
it completes.
After the change, we will handle NCELLMEAS notification whether or
not lte_lc_neighbor_cell_measurement() had been called or not.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>