applications: matter: Fix for handling status in Matter Bridge
Currently, the Matter Bridge cannot read the current LED state
from the LBS characteristic. After losing the connection between
the Bridge and the BLE LBS device we lost information about the
current LED state.

To avoid losing the LED state synchronization, when the Matter
Bridge retrieves the connection to a Bluetooth LE LBS device we
should set the current value of LED on the BLE device. We should
focus on two scenarios:

1. If The Bridge has not been rebooted and the connection is
retrieved, then the state of the LED that is kept in RAM should
be set on the LBS device.
2. The Bridge has been rebooted and it restored all connections
to BLE devices, and then we should force the initial LED state
on the BLE device.

Another change is regarding ESP subscriptions. We cannot
subscribe to the GATT in the Init method because we do not have
an active connection. We should subscribe after checking all GATT
descriptors and unsubscribe when the lost connection is detected
and stop the timer.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>