applications: asset_tracker_v2: Default cellular timeout 11s -> 30s
Neighbor cell measurement timeout has been set to 11s by default
due to the time it took before GCI search was available.
Now that it's available and Location library is using it, this
limitation is unnecessary and we should take advantage of it.
Location library default of 30s will be used in ATv2 unless otherwise
configured in ATv2 side.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>