cmake: adjust PM related CMake variables to global properties
Upstream Zephyr PR#57884 adjusted the inclusion of sysbuild modules
so that these are now sourced in a different scope.

This cause PM related variables to not be available when running
partition manager CMake code as part of sysbuild.

PM related variables are therefore changed to global properties, making
the partition manager related code more robust.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>