bl_storage: Fix nv_counters bug with FIH
The NRF non-volatile counters function boot_nv_security_counter_get
is called by MCUBOOT using the macro FIH_CALL(). When the fault
injection protection is enabled this macro increases a counter
in MCUBOOT which is expected to be the same before and after the
call to the function. The counter is increased by FIH_CALL() and
is expected to get decreased by the FIH_RET() macro. This means
that all the functions which are called by FIH_CALL() need to
return with FIH_RET(), otherwise the fault injection protection
will stop the execution.

This change fixes a bug with our implementation of nv_counters which
didn't return FIH_RET(). This had as a consequence that when the
hardware downgrade prevention and the fault injection protection
was enabled in MCUBOOT:
CONFIG_MCUBOOT_HARDWARE_DOWNGRADE_PREVENTION=y
CONFIG_BOOT_FIH_PROFILE_LOW=y

MCUBOOT could not boot anymore.

Ref: NCSDK-23761

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>