lib: location: Ignore timeout if another event processing ongoing
It's possible that we get GNSS fix and when we are sending an event
about it to the application, location request timeout expires.
When an event is indicated to the location library core,
a check is made whether an event processing is already ongoing.
And if there is, new event is ignored.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>