manifest: pull Matter fix for IPv6 prefixes count
Zephyr ipv6_nbr implementation requires the given
address to have a matching prefix set on the
interface. Otherwise, the default router is used
for sending neighbor advertisement and as a result,
in case there are multiple routers in the network,
the packet can be sent to the invalid interface
(not the one which issued neighbor solicitation).

The detailed analysis of the issue can be found in KRKNWK-17318.

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>