bootloader: Fix issue with public keys when MCUboot is not from source
When MCUboot is provided as a hex file, the b0 signing code is not
enabled, since a signed edition of MCUboot can be provided, but
this code also populates the right public key into the public key list
in the provision page, so when it is skipped, the provision page is
missing a public key.

This splits the sign.cmake file into sign_keys.cmake and sign.cmake
and only skips the actual signing, not the preparation of the keys,
in the case that MCUboot is not build from source.

Jira: NCSIDB-883

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>