samples: central_uart: Add support for unsecure connection
The peripheral_uart supports the Kconfig option BT_NUS_SECURITY_ENABLED
which allows the sample application to be built with or without support
for a secure connection.

This option is not present in the central_uart sample, it
unconditionally enforces a secure connection.

If the peripheral_uart sample application is built without support for a
secure connection, this leads to the following error and warning
messages in the log file for the central_uart sample application when a
connection is established:

<err> bt_smp: pairing failed (peer reason 0x5)
<wrn> central_uart:  Security failed: xx:xx:xx:xx:xx:xx (random)
      level 1 err 5
<wrn> central_uart: Pairing failed conn: xx:xx:xx:xx:xx:xx (random),
      reason 5

This patch adds BT_NUS_SECURITY_ENABLED to the central_uart sample
application similarly to the peripheral_uart sample application, but the
default behaviour when building the sample application is unchanged.

Signed-off-by: Michael Hunold <michael.hunold@eu.panasonic.com>