samples: central_uart: Make UART_RX_TIMEOUT match peripheral_uart
The peripheral_uart and central_uart sample applications use almost
identical code for the UART handling.

But one difference is the usage of the UART_RX_TIMEOUT value. It is used
in uart_rx_enable() and is expected to be a value in microseconds.

The peripheral_uart sample application uses a value of 50000 for
UART_RX_TIMEOUT which corresponds to a timeout of 50ms, which is
pretty reasonable for a timeout value for the UART.

Due to whatever reasons the central_uart application uses a value of 50
for UART_RX_TIMEOUT which corresponds to a timeout of just 50
microseconds, which looks unreasonable.

The assumption here now is that this was simply an oversight and really
50000 should have been used right from the beginning.

This patch fixes this and makes the handling of peripheral_uart and
central_uart identical when it comes to UART_RX_TIMEOUT.

Signed-off-by: Michael Hunold <michael.hunold@eu.panasonic.com>