ppi_trace: add ppi_trace_dppi_ch_trace function
For current ppi_trace_config it is impossible to trace events for
which PUBLISH register is overwritten in run-time after
ppi_trace_config is called. This function allows also to trace just
one event and does not allow to trace many-to-one topologies
which are possible with use of DPPIs.

To overcome above limitations ppi_trace_dppi_ch_trace is added.
In contrast to ppi_trace_config the function ppi_trace_dppi_ch_trace
focuses on tracing rather the DPPI channel leaving the preparation
of event network publishing to it to a user.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>
(cherry picked from commit 503f8e4cad2467769a62a92c1be001bb0f8063e2)