tests: fprotect: negative: Fix failing test
The test started failing after an upmerge. It was likely because of the
change that makes PICOLIB the default library. The test failed because
memcpy didn't trigger a bus fault. Probably because of some compiler
optimization. Switching to MINIMAL_LIBC works, but this fix doesn't
require that.

Rearrange the tests and make key variables volatile to be more robust
against optimization.

Also properly check the output of fprotect_is_protected()

Jira: NCSDK-23967

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>