modules: mcuboot: Invert dependency of MCUboot mode
If we enabled building images for secondary slot when MCUboot
was in Direct-XIP mode using CONFIG_BOOT_BUILD_DIRECT_XIP_VARIANT
we set automatically MCUboot mode of operation to direct-xip.

It makes more sense to explicitly set the MCUboot mode of operation
and enable building of the secodary slot automatically.
Also as there is a support for the direct-xip with revert mechanism,
we don't have to overwrite set MCUboot mode of operations after
enabling building of the secondary slot.

Aligned nrf_desktop, direct-xip test and peripheral_lbs.

Jira: NCSDK-24079

Signed-off-by: Mateusz Kapala <mateusz.kapala@nordicsemi.no>