drivers: wifi: Update RPU patch to rev#ca0df346294
Hardware was seen not resetting the MIC error bit before processing the
plain packet. Hence when a MIC error occurs due to CRC failure, the MIC
error bit is set, which is not reset at a later stage. This causes the
firmware to drop genuine packets on the account of MIC error.

Signed-off-by: Divya Sai Keerthi Tiruvakadu <divya.tiruvakadu@nordicsemi.no>