drivers: wifi: Fix crash during de-init of the interface
During interface de-initialization WPA supplicant thread is still
running and issuing calls to the driver, esp. set_supp_port (802.1x port
status changing during connection and disconnection), so, for now use
the network interface lock to synchronize with data path. This fixes
NRF7X-93 only.

But not only set_supp_port, this is a generic race b/w control and data
path and needs a wider fix which will be addressed in SHEL-2077.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>