applications: serial_lte_modem: change Zephyr settings backend (FCB to NVS)
With this change the settings previously stored are lost.

This required changes to the FOTA logic such that application
updates do not require settings to store the state.
This is achieved with the use of mcuboot_swap_type().

The same changes are made to modem delta updates as a callback
gets called when the modem library is initialized after such
an update.

One FOTA setting remains to keep track of ongoing full modem
updates because nothing otherwise indicates them if the device
is reset. This is done to conserve the possibility to apply
a full modem update with a device reset.

Also, the lack of support of full modem update on the
Thingy:91 is made explicit in the FOTA documentation
and guarded against by CONFIG_SLM_FULL_FOTA.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>