secure_storage: Add encrypted secure storage subsys
This adds a sub-system to provide a secure storage solution.
It implements the PSA Internal Trusted Storage and PSA Protected
Storage API's
The system is written in modular way to allow different backends to
encrypt and authenticate the assets written but also how the assets
will be written to non-volatile memory.

The default backends will use the settings sub-system to store the
assets and ChaCha20-Poly1305 to encrypt and authenticate them.
By default the PSA Crypto API's are used to perform the AEAD operations.
Additionally a mbedTLS implementation is provided.
Alternatively a custom implementation can be added to use the platform
specific crypto accelerator directly.

Ref: NCSDK-22876
Ref: NCSDK-21874

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>