doc: migrate nrfxlib doc to sdk-nrf
This pull request moves all nrfxlib documentation
into the NCS docset in an attempt to remove the
nrfxlib docset completely.
Doc source files for nrfxlib will no longer reside with the code.

Main reason for doing so is faster CI.
Currently, we need to build nrfxlib docs twice because of
interlinking between nrfxlib and nrf docs.
It's also a better experience for the user if all
information can be found within one doc system.

Signed-off-by: Bartosz Gentkowski <bartosz.gentkowski@nordicsemi.no>