nrf_security: Align AEAD setup of operation
In all the functions apart of the AEAD ones
the operation ID is set only if the setup call
of the driver returns success. This makes sure
that any subsequent calls (update/finish) will
not even reach the driver call if the setup was
not succesful. This aligns the AEAD to follow
the same pattern.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>