drivers: wifi: Add support to flash the patch automatically
If the storing of nRF70 FW patch in external flash (non-XIP) is enabled,
user needs to two steps, flash the patch to external flash and regular
Zephyr image flash.

Add support to automate this process either using DTS or PM:

0. Convert the patch binary to hex using "intelhex" python package.
1. Generate a merged hex file directly if using DTS or patch hex file
when using PM and configure PM to merge this.
2. Set the merged image as default hex file for the flash if using DTS,
if using PM this is done automatically.

This way, without any user involvement we can automate the process.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>