tfm: Use NS fault handler service to dump fault information in nonsecure
Use the TF-M non-secure fault service and take the context information
and re-use the zephyr fault handling.
The zephyr z_arm_fault function cannot be used as this one accesses
HW registers directly. Instead we mirror the functionality but access
the context state instead. Some simplifications can be made as the
possible states and faults are limited.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>