tests: lib: location: Improvements to make debugging easier
Small improvements to be able to get more information when tests fail.
Adding asserts on k_sem_take() return values.
Adding 1ms sleep to the end of all functions waiting for events.

Still not moving TEST_ASSERT_*'s out of location_event_handler()
as mentioned in Jira NCSDK-21340.
Let's see if this helps or if that has to be done too.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>