kconfig: fw_info: let entry be a menuconfig, rework FW_INFO_API
Let fw_info entry FW_INFO be a menuconfig, so that its related
entries can expand nicely into a menu.

The dependency on FW_INFO || FW_INFO_API on the Kconfig entries
for this library prevented them from expanding into the menu.
Rework FW_INFO_API so that it depends on FW_INFO.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>