lib: nrf_modem: lte_net_if: remove driver _DOWN flags
This driver has a flag controlling the behavior of net_if_down().
One value of the flag lets net_if_down() bring the interface down,
the other lets net_if_down() disconnect from the network.

This flag can be set at run-time.

The application has at its disposal a dedicated API to bring
the interface down, i.e. net_if_down(), and a dedicated API
to disconnect, i.e. conn_mgr_disconnect().

Thus there is not need to change the behavior of net_if_down()
based on a run-time flag to perform another API's operation.

At run-time, the application shall call one of the two functions
to achieve the desired behavior.

However, the application may want to align net_if_down() behavior
with that of other interfaces for the purposes of behaving similarly
when net_if_down() is called outside of the application, e.g.
in a library.

For that purpose, a build time configuration option to control
the behavior of net_if_down() is desirable.

This configuration option exists and remains untouched.

This patch removes the _NET_IF_DOWN flag support from
the lte_net_if network interface driver.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>