Applications: nrf5340_audio: synchronize L and R with SDC.
Use timestamps in SDUs to synchronize the left and right channels. The approach to
synchronize the left and right channels are the following. If two
channels send their first SDU simultaneously, it is without a timestamp
as the controller cannot provide a timestamp value using
read_iso_tx_sync until after the first SDU has been sent.
The first two SDUs will be given a timestamp by the controller. This
timestamp is based on time of arrival. Unfortunate timings can lead to
the two timestamps belonging to different CIG reference points. In this
situation, read_iso_tx_sync will return different tx timestamps for the
two channels. The channel that is behind must use the timestamp of the
channel that is in front in time to catch up and synchronize.