applications: serial_lte_modem: allow modem AT command interception
For the interception to catch all AT commands, they all need
to be sent through nrf_modem_at_cmd() because this is the only
function covered by the modem library's callback feature.

This was already the case for AT commands received externally,
but not for AT commands sent from the SLM code.

This implements alternative functions to nrf_modem_at_printf()
and nrf_modem_at_scanf() as well as trap macros to prevent
their use.

Also, because the interception is case-sensitive, all the
AT commands sent are now all-uppercase.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>