doc: build files: move the list to new page
Moved the reference list of output build files to a new page
in the Configuration and building section. Edited other pages for links.

Removed Building FOTA images section and edited info in tables and
cross-links.

Added link to MCUboot build output files in Adding MCUboot as an
upgradable bootloader.

NCSDK-22324 and NCSIDB-964.

Signed-off-by: Grzegorz Ferenc <Grzegorz.Ferenc@nordicsemi.no>