samples: wifi: shutdown: Add an option to disable auto-start
In case the interface auto-start is disabled, Wi-Fi driver initializes
and waits for user to bring the interface up. This options is useful for
customers that need finer control over when Wi-Fi driver starts comms to
nRF70 chipset (depends on driver).

The sample doesn't attempt to bring the Wi-Fi until user presses the
button1.

Implements NRF7X-109.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>