scripts: Add memory thresholds' list
This commit is part of a solution addressing developers' need
to be aware of increases of memory footprints of applications
under their maintainership.
The list added in this commit stores information about configurations,
whose memory footprints should be monitored in the CI.
Additionally, this list contains thresholds passing of which will
trigger an automated warning on PRs introducing such changes.
This dictionary is used as an input by a script computing memory
increases, called by the CI after the twister stage.

jira: NCSDK-23524

Signed-off-by: Katarzyna Giądła <katarzyna.giadla@nordicsemi.no>