ext: oberon: Align with the Oberon deliverables
This aligns the Oberon PSA core and driver code with
the Oberon deliverables.
All the modified files don't have any functional change,
it's only code structure changes.

There are two new files, the sha256.h and the
memory_buffer_alloc.h. These were added to align
with the deliverable and they should not make any difference
since they are identical to the same files in sdk-mbedtls.

Ref: NCSDK-24707

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>