wifi: l2ac: Fix connection persistence
If the persistence is enabled and In case WPA supplicant disconnects for
seem reason e.g., AP de-auth due to wrong PSK, then L2AC should not
issue a disconnect until the connection timeout is hit.

But default timeout is 0, so, we should check before we reschedule the
timer to avoid disconnecting immediately, as we have a separate API
which properly does this, use it.

Fixes SHEL-2249.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>