doc: debugging: Document how to use Monitor-mode debugging
Monitor-mode debugging greatly enhances the debugging experience of
Bluetooth applications and other applications where timing-critical
code needs to continue execution when paused on a break-point.

This commit adds a section about this topic under the section
`Testing and debugging an application`. This section should make
it more visible to users how Monitor-mode debugging can be used
in the context of the nRF Connect SDK with the tools provided
by us.

References to documentation provided by the Zephyr project and
SEGGER are added for users that want to get a deeper insight on
how this works and how it can be used.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>