applications: Fixed DFU over SMP for Matter bridge
The DFU over BT SMP was sometimes failing becase advertising
restart was not properly handled in the Matter BLEManager, due to
the CHIP_DEVICE_CONFIG_CHIPOBLE_SINGLE_CONNECTION flag.
Additionally it was required to pull changes in Matter code that
removes collissions in handling peripheral and central related
BT callbacks.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>