lib: location: Skip GNSS if not in RRC idle
If timeout for waiting for RRC idle mode expires
and we are still in RRC connected mode, let's not try to get GNSS fix.
It will only spend time and power without getting a fix.
We also spend the location request timer and fallbacks are not even
tried.

In addition, lowering the waiting time for getting to RRC idle mode
and PSM from 5mins to 2mins.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>