net: lib: nrf_cloud: fix FOTA settings handling
When the base nRF Cloud MQTT settings are loaded, the FOTA
settings are also loaded. Previously, the pointer was not
yet set, resulting in an assertion.
Updated to ignore if the pointer was not set.
Removed duplicate settings loading in nrf_cloud_fota.
Updated nrf_cloud_fota_settings_load to always load the
settings instead of just once.
IRIS-7756

Signed-off-by: Justin Morton <justin.morton@nordicsemi.no>