mpsl: fem: require PINCTRL for nRF21540 GPIO+SPI driver
The current nRF21540 GPIO+SPI driver implicitly depends on PINCTRL and
relies on the end user application to configure it appropriately. This
commit makes the dependency explicit: the nRF21540 GPIO+SPI driver
selects CONFIG_PINCTRL and throws a build error if that Kconfig option
is not set. Code executed when pin controller was disabled was removed.

Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>