wfa-qt: Address association request capture failure
The `enable_network` command is part of the configure API, and it
is resulting in advance connection and causing QT tool failures
in capturing association frames after sending association command.

To address the issue, move the `enable_network` command from the
configure API to the association API to ensure that the connection
is established only after receiving the association request command.

Fixes SHEL-1861 and SHEL-1862

Signed-off-by: Triveni Danda <triveni.danda@nordicsemi.no>