secure_storage: Fix counter bug
The counter used as a nonce for the secure storage was
16 bytes but the implementation on the AEAD backend
expected a 12 byte nonce. This corrupted the key material
when the nonce was written.
This fixes the issue by making the counter to be 12 bytes
as expected by the AEAD backend. It also ensures the AEAD
backend and the counter have the same length and fails at
runtime without corrupting any buffers if they differ.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>