applications: serial_lte_modem: second half implementation of PPP
This adds auto-start functionality so that the PPP link
goes up/down when the network connection goes up/down.

Also, auto-restart is performed when the PPP peer
disconnects so that reconnection can happen successfully.

Documentation for the #XPPP AT command is added. The rest
is not yet documented as this is still a work in progress.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>