secure_storage: Ensure counter is initialized
When we try to get a counter we need to be sure that
the counter is initialized before use. Otherwise we might
end up with an all zero counter if a user of this starts
early in the boot process.
This change ensure that the counter is initialized before
use.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>