manifest: Update Matter revision to pull PSA changes
We need a mechanism to move the DAC private key from
the factory data set to PSA ITS NVM storage during the
first boot of the device.

This feature is available only if the CONFIG_CHIP_CRYPTO_PSA config
is set to y.

You can enable removing the DAC private key from the factory data
partition after the migration by setting the
:kconfig:option:`CONFIG_CHIP_CRYPTO_PSA_MIGRATE_DAC_PRIV_KEY`
Kconfig option to ``y``

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>