manifest: Update Matter SDK revision to pull new crypto changes
In this commit added a mechanism that migrates all existing
operational keys from mbedTLS-related settings storage to the
PSA ITS secure storage.

After the migration the entries in the settings storage will not
be avaliable anymore.

From NCS 2.6.0 the Matter supports Arm PSA Crypto API by default,
and mbedTLS is deprecated.

We should not use CommonServerInitParams to initialize PSA crypto
and we should do it directly in the Application instead of
Chip::Server. To do it we moved PSA init and declaration from
the Chip::Server to the Nordic's Matter Init.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>