samples: bluetooth: throughput: Negotiate con params before packet len
This commit changes the order of the connection configuration.
Negotiate the connection params with higher connection latency
before changing the packet len.
This way we may be sure that the connection would not break
due to timeout after the high packet len was set that cannot
fit into configured connection latency.

JIRA: NCSDK-25371

Signed-off-by: Radoslaw Koppel <radoslaw.koppel@nordicsemi.no>