nrf_security: Separate PSA_WANT configuration for Mbed TLS config
Due to the split build of TF-M we need to pass the PSA_WANT
configuration to the Secure and Non-Secure image.
Before TF-M was building both images in the same process but now the
Non-Secure image is build separately. But for some images like tests
we need to know the PSA configuration for both images, but so far the
PSA_WANT configuration was ignored if legacy was used in the
Non-Secure.

This commit places the setting the PSA_WANT configurations in a
separate cmake file so it can be used by legacy and psa.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>