samples: lwm2m-client: Change eDRX cycle to 5.12s
When we have shortest eDRX cycle, it is most responsive
so there would be less likely CoAP retransmissions.

This is good setting for development phase, as it is most
compatible with default settings of common servers.

For power saving, developer might choose to go higher,
but it may cause more CoAP packet losses.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>