net: lib: nrf_cloud_coap: Fix improper coap_client init
If there is a delay between calling coap_client_init() and
performing a transfer, during which another thread calls
coap_client_init() and does a transfer, the coap_client library
gets confused because each struct coap_client has fd=0. This
causes the first one initialized to be given received data,
while the second one, which asked for the data, does not get it.

Initialize the fd field to -1 so the coap_client library
does not send it the wrong data.

Jira: IRIS-8121

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>