net: lib: nrf_provisioning: Clear client struct fd on close
When the library closes the socket, it leaves the
fd field in the struct coap_client set to the old fd.

This can cause the coap_client library to send data
intended for another thread (which was given the same fd
when it subsequently opened a socket) back to this thread,
causing errors in both threads.

Jira: IRIS-8121

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>