net: lib: nrf_provisioning: Print attestation token
When the device cannot connect to the provisioning
service due to being unauthorized, this means the
user must claim the device on nrfcloud. Help the user
by printing a message to that effect, and display
the attestation token so they can claim the device.

Add overlay for CoAP provisioning. Rename HTTP provisioning
overlay to be consistent.

Add workaround to nrf_cloud_coap_transport for file
descriptor reuse bug in coap_client.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>