mpsl: Define CONFIG_MPSL_PIN_DEBUG
Toggling pins on radio events is something that can be useful
for debugging purposes. By defining Kconfig entries,
this becomes simpler to setup.

The default configuration will setup two pins to toggle
on (RADIO->EVENTS_READY, RADIO->EVENTS_DISABLED) and
(RADIO->EVENTS_ADDRESS, RADIO->EVENTS_END).

The implementation may require code to be run on both the
application core and radio core.
As Kconfig entries are not propagated to child images,
this entry needs to be set for both cores.

For example like this:
```
west build -b nrf5340dk_nrf5340_cpuapp --
-DCONFIG_MPSL_PIN_DEBUG=y -Dhci_ipc_CONFIG_MPSL_PIN_DEBUG=y
```

The entry is marked as experimental because the implementation
is not necessarily stable nor tested properly.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>