encrypted_storage: Rename from old secure storage subsystem
The name secure storage is not optimal as the word secure comes with a
lot of different assumptions, that might not be given for the assets
stored with that subsystem. Therefore we rename it before the 2.6.0
release.

Encrypted storage is a more direct description of what it does, even
though due to using a AEAD scheme it provides more than just encryption,
it is the bare minimum.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>