samples: cellular: modem_shell: Add sock send --wait_ack
Add an option to sock send to specify MSG_WAITACK. This will request
a blocking send operation until the data has been sent on-air and
acknowledged by the peer if required by the network protocol.

Set requested blocking mode for a send with --data option because
this will affect --wait_ack.

Increased time precision from 2 to 3 decimals in throughput summary
to get milliseconds.

Signed-off-by: Stig Bjørlykke <stig.bjorlykke@nordicsemi.no>