application: serial_lte_modem: integrate CMUX with PPP
The UART-based PPP driver is replaced by
the PPP module of the modem subsystem.

When both CMUX and PPP are enabled, CMUX automatically configures
a second channel for PPP which can only be used through it.

Thus, CMUX should be started before PPP, which means before activating
LTE as PPP starts automatically when the LTE link goes up. However,
SLM tolerates that CMUX is started only after LTE and PPP are up.

As of now, upstream Zephyr commits are needed for this to work properly.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>