debug: etb_trace: Start trace and debug clock on idle exit
This patch fixes an issue where the trace and debug clock was
not restarted when exiting idle state, which caused the
ETB trace collection to hang indefinitely.
The issue could be observed when CONFIG_ETB_TRACE_LOW_POWER was
enabled and after CONFIG_PM was no longer enabled for nRF targets.

Fixes CIA-1064

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>