Applications: nrf5340_audio: Use VS TX sync command when possible
The Read TX sync command is used to synchronize data transferred
from the application to the network core. It is used
to figure out if the data was sent to early or too late.

In this context, it is the assigned sync ref that matters, it is the
sync ref that determines when a packet will be sent on air.

The HCI spec defined TX sync command returns a CIG/BIG reference anchor
point. This may or may not be aligned with the corresponding sync ref.
Therefore, we use a vendor specific command instead.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>