Bluetooth: Mesh: Add a note about warning caused by fixed passkey option
When `CONFIG_BT_FIXED_PASSKEY` is enabled, it generates the warning that
this option should not be used for production. This should not be a
problem as far as the passkey is generated using secure RNG, generated
on every pairing request and the passkey is transferred over mesh. These
all correct for the LE Pairing Responder model use-case. Therefore,
adding a note that this warning can be disregarded.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>