lib: supl: Implement workaround for a SUPL client library bug
Implemented a workaround for a bug in the SUPL client library.
The library causes a crash if the status field in the lte_params
structure is set to false. As a workaround, the status field is
now always set to true. If LTE registration information is not
available, all parameters in the structure are zeroes, which
is handled gracefully by the library.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>