net: lwm2m_client_utils: Set RAI_ONGOIN if we receive cmd
Sometimes LwM2M engine is already indicating NO_DATA but
we are not yet on QUEUE_RX_OFF state.
If we receive CoAP packet during this period, it means that
server have something for us.
But as the callback does not get any indication on RX side
we only know that we have received CoAP traffic if the state
transitions from NO_DATA to LAST.
When this transition happen, stop the RAI indications until
we are in silent period for a while (QUEUE_RX_OFF).

This fix a behaviour seen on Nb-IoT where RRC connection
is released and started between each CoaP GET/Ack pair.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>