applications: serial_lte_modem: integrate with Zephyr generic modem driver
This makes the changes necessary for SLM to be compatible
with Zephyr's generic modem driver.

The CMUX channels are now run-time configurable such that
both AT commands and PPP can be on any of the two DLCIs.
This is achieved by explicitly binding the AT channel to
a particular DLCI, and letting PPP take the other one.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>