samples: matter: Made IsGroup binding parameter optional
The BindingData structure requires setting IsGroup argument, which
can be used to apply specific command to only a part of binding
entries (groups only or unicast only). This can be useful, but
the default behavior should be sending command to all targets
present in the binding table.

Changed IsGroup field to be optional. If it's not present the
command will be sent to all targets. Otherwise the old filtering
scheme will be used.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>