hostap: Delete untested WPA_SUPP_CRYPTO_LEGACY_BUILTIN option
Delete the option WPA_SUPP_CRYPTO_LEGACY_BUILTIN.

It is is not documented outside of Kconfig.

Not used by any samples and is therefore subject to bit rot.

And is bad practice to use, as we have WPA_SUPP_CRYPTO_LEGACY which I
believe will address the same set of use-cases.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>