tests: fota_download: Fix mocking of tfm_platform_s0_active
Due to split build feature of TF-M 2.0.0 our old trick of mocking the
tfm_platform_s0_active(...) doesn't work anymore.

Removing stuff from the old mocking, and adding a Kconfig that allows to
not include the mocked functions.

Also removing the mock of tfm_platform_ioctl() as it is not needed
anymore.

Ref: NCSDK-26150

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>