drivers: wifi: Increase the interface name length
commit 23bf5c6f411("drivers: wifi: Use DT macro for defining device")
renamed interface name from wlan0->nordic_wlan0, but as the default
maximum interface length is only 8, this causes issues while trying to
remove the interface in the WPA supplicant as we only copy the length
defined by this configuration in remove, but use full interface name in
add.

Fixes SHEL-2463.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>