location: Fix CMock unit test
Now that we are usign DT macros for nRF70 Wi-Fi, the CMock which runs on
native posix board fails as there is no Wi-Fi interface on native posix
and no mocking for DT APIs.

To solve this add a board specific overlay with the Wi-Fi DT label
defined and also add a dummy Wi-Fi device instance.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>
Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>