modules: trusted-firmware-m: Add NFC pins option
Add option for configuring the NFC pins in TF-M.

The option CONFIG_NFCT_PINS_AS_GPIOS in Zephyr
is being deprecated and it will be removed soon in
favor of DTS configuration for this.

TF-M doesn't support DTS and thus we cannot use
the DTS option directly.

This adds a separate option for TF-M which will not
be deprecated and make use of the of generated Kconfig
from DTS to set it.

Ref: NCSDK-24437

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>