samples: wifi: Increase polling sockets
Recent changes to Zperf caused an increase in maximum polling sockets as
it now uses socket services (TCP-6, UDP-2) for it's receiver, and due to
a but in the socket services, there is a array overflow causing a crash
and corruption in BLE Co-ex sample.

Increase the sockets to fix the warning at the boot and the corruption,
the proper fix to Zephyr socket services will be done separately.

Fixes SHEL-2530.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>