samples: wifi: Fix SML experimental update for nRF7000 and nRF7001
The idea was to skip twister for these variants to save CI time, but
still add an entry for SML. But SML in current state needs access to
"build.log" and ".config" build artifacts, so, the experimental SML
isn't updated properly.

Till we figure out a way to get experimental status without build
artifacts, enable these experimental features for boards nRF7000 and
nRF7001 in the twister.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>