hids: Propagate errors in output and feature report handlers
Custom feature reports are often (ab)used in custom HID devices to
implement arbitrary command/response handling and as such often
require an ability to reject a given payload and signal an error
condition. USB HID implementation in Zephyr allows that, so implement
a similar capability in HIDS to match.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>