nrf_rpc_uart: transport refactoring
nrf_rpc_uart: fix hdlc encoding

The code couldn't properly encode and decode a frame that
contained 7d byte (delimiter). Change the encoder to XOR
an escaped byte with 0x20 like in the original HDLC spec
and fix the decoder to properly handle escaped bytes.

Also, increase the max frame size to 1536 to be able to
send large IPv6 packets over RPC and move the frame buffer
to the transport context.

nrf_rpc_uart: fix RX buffer usage

Use a different buffer for the RX ring buffer and for
reading bytes from UART peripherals. This fixes the issue
with losing first byte on one side of the communication
after reboot.

Signed-off-by: Przemyslaw Bida <przemyslaw.bida@nordicsemi.no>
Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>