nrf_rpc_ipc: Lazy waiting for endpoint bond
This commit removes the waiting for the remote cpu while initialization
is done.
The waiting is moved to a sending function.
This way we would be able to initialize multiple IPC endpoints
also in other modules and wait for the bond to happen only
if we need to transfer anything.

JIRA: NCSDK-26208

Signed-off-by: Radoslaw Koppel <radoslaw.koppel@nordicsemi.no>