samples: ipc_service: Let IPC backends be enabled based on DTS
Kconfig options for IPC service backends are enabled by default when
an IPC node with the corresponding compatible property is enabled in
DTS, so there is no need to enable those options in the application
configuration files. And doing so may lead to Kconfig warnings when
the enabled option does not match the IPC node compatible property
(this was the case for nRF54H20, where the compatible property for
the cpuapp_cpurad_ipc node is "zephyr,ipc-icmsg-me-initiator" or
"zephyr,ipc-icmsg-me-follower" but the enabled Kconfig option was
IPC_SERVICE_BACKEND_ICMSG).

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>