samples: dtm: align to fem_al tx_power_control
In the fem_al the "gain" of FEM is changed to "tx_power_control".
This commit aligns to these changes.

Due to these changes, the DTM command
FEM_GAIN_SET allowing to set "gain" of the FEM makes no more sense
and is turned into:
FEM_TX_POWER_CONTROL_SET allowing to set the FEM tx power control
which is FEM type-specific.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>