tests: tfm_psa_test: Fix sample configuration for 54L
The _ns.conf file is no longer necessary.

We keep the overlay file because we want to use a different overlay
file for 54L than in old platforms.

old platforms are picking up the app.overlay file in the app
directory, but it refers to uart0 which is not present on 54L.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>