tfm: rewrite 54L TF-M CMake support
We move most of the 54L TF-M CMake support to TF-M.

This was done to make 54L align as much as possible with other
platforms.

This makes comparing different platforms easier as all files are in
similarly placed directories.

We also port the remaining existing CMake code to the new TF-M
revision.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>