Bluetooth: Mesh: Remove unneeded dependency from metadata extraction
Make it possible to extract metadata at build time even if
BT_MESH_DFU_METADATA is not enabled in the image. There is no need to
have this enabled in the new firmware for it to make sense, as long as
it is enable in the firmware currently running on the target. Removing
the dependency allows things like removing Mesh DFU support by
performing a Mesh DFU update, or changing metadata schemes by performing
a DFU update.

Also remove the default y for BT_MESH_DFU_CLI. It makes no sense to
enable it by default for distributor images, unless they support
self-update, in which case they would also enable BT_MESH_DFU_SRV.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>