samples: wifi: throughput: add CONFIG_NET_SOCKETS_POLL_MAX entry
Add CONFIG_NET_SOCKETS_POLL_MAX entry and set entry for poll call
to 10 for throughput sample.

It is seen that the current througput build requires POLL_MAX to
be configured to atleast 9. however, the default seems set to 7.
Setting the POLL entries to 1 greater than the max poll entries
required for build similar to other samples for Wi-Fi

Fixes SHEL-2602

Signed-off-by: Vivekananda Uppunda <vivekananda.uppunda@nordicsemi.no>