partition_manager: cmake/python: Add DTS as input for partition_id
Sometimes we have issues where the FLASH_AREA_ID used is defined from
DTS. This tries to remedie these types of issue to ensure that the
FLASH_AREA_IDs for `fixed-partitions` are the same across DTS generated
header and Partition Mananger generated header.

This way if you select a partition/flash_area from the DTS define it
would map to the correct FLASH_AREA which was defined by partition
manager resolving in less conflicts between DTS and PM.

The goal of this patch is to fix the issue regarding using
`boot_write_img_confirmed()` function in to confirm MCUBoot images.

Fixes: NCSDK-26693

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>