cmake: multi_image: Forward CROSS_COMPILE
According to the Zephyr documentation, ZEPHYR_TOOLCHAIN_VARIANT take
the value of `cross-compile`, in which case, the additional variable
CROSS_COMPILE has to be set to a common path prefix. If this
CROSS_COMPILE is not forwarded to the build of sub images, the build
will fail.

(https://docs.zephyrproject.org/latest/develop/toolchains/other_x_compilers.html)

Signed-off-by: Torsten Robitzki <Torsten@Robitzki.de>