bluetooth: controller: Put ECDH driver in the MPSL work queue by default
Puts the SDC ECDH driver in the MPSL work queue by default, unless
BT_SMP is selected. It will also be put in the MPSL work queue by
default when split host/controller build configurations are used.

It's expected that many use-cases do not use this module,
and that many applications that do use it, do not necessarily need to
allocate a dedicated thread for its processing. Placing the ECDH work
in the MPSL work queue by default saves considerable RAM for those
use-cases.

We estimate that the ECDH driver processing can take time in the dozens
of milliseconds (possibly more) in the worst case.

Signed-off-by: Olivier Lesage <olivier.lesage@nordicsemi.no>