modules: hostap: Increase the stack size for net_mgmt
Due to recent changes in hostap to move few buffers from heap->stack (to
improve the reliability of allocations) this has increased the stack
size and with another recent change to increase the max polling sockets
this puts over the 4096 stack size for net_mgmt causing stack overflow.

Increase the stack size for net_mgmt to fix SHEL-2640 and SHEL-2676.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>