Bluetooth: Host: introduce priority sets
In order to better adapt to the needs of the application, add a host API
that allows to change the priorities in the controller.
These configuration sets are tested internally and have a specific
purpose.

Initially, add a configuration which gives more priority to the
initiator, in order to improve connection setup latency.

Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>