partition_manager: Allow to disable PM for a project
The commit allows to control whether Partition Manager
is enabled for build of selected app or not.

Note that when selecting CONFIG_PRE_SELECT_PARTITION_MANAGER=n
it should happen either in project file or directly in command
line for the first time project is build, otherwise the
Partition Manager will be enabled by default and it will affect
many configuration options, something that can not be reversed
easily.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>