partition_mananger: Change TF-M storage placement
In nRF5340 and nRF9160 the TF-M storage is placed after the
the application itself. This results to our secure/non-secure
partitions to be split like this:
TF-M image          -> secure
Application image   -> non secure
TF-M storage        -> secure
Application storage -> non secure

This doesn't work well with nRF54L15 since
we have a limited number (4) of SAU regions to configure.

For nRF54L15 this change modifies the partition locations
so that the TF-M storage is always placed before the
Application image.
This will allow us to reduce the partition secure/non-secure
configuration in the SAU since it will allow us to configure
the TF-M storage and the TF-M image as one partition:
TF-M storage        -> secure
TF-M image          -> secure
Application image   -> non secure
Application storage -> non secure

Note that when MCUBOOT is not used we place the TF-M image
first followed by the TF-M storage, but this doesn't affect the
configuration.

When MCUBOOT is enabled the TF-M storage partition is
placed before the MCUBOOT primary partition in order to
avoid being affected by updates.

Ref: NCSDK-24863
Ref: NCSDK-27354

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>