54l: ns: Fix the ns board configuration wrt. grtc IRQs
nrfx has hardcoded in nrf_grtc.h that ns should use GRTC_1, which
corresponds to irq 227. And that secure should use 228.

So we change the IRQ in the non-secure board from 228 to 227.

Also, add 229 which is used by mpsl, this change corresponds to a
patch that already exists in the secure board's DT.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>