modules: tf-m: nRF54L15 fix TF-M+MCUBOOT alignment
Align the TF-M parititions for nRF54L15 so that the
non-secure application starts with the SPU based
alignment which is currently in place. Note that the
SPU alignment is only used as a workaround now,
the nRF54L will need to configure the non-secure application
based on the MPC and SAU alignment requirements.

In nRF54L15 there are two main things that changed regarding the
partition placement and size:
1) The TF-M storage partitions (ITS/PS/OTP) are always placed
   before the non-secure app. Depending on the configuration
   they are placed either before the TF-M image itself
   (non-MCUBOOT case) or after the TF-M image and before
   the non-secure app (when no bootloaders are used).
2) The MCUBOOT padding partition for nRF54L15 is increased
   because the vector table needs to align to an address of
   multiple of 0x800.

Ref: NCSDK-24863
Ref: NCSDK-27354

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>