scripts: Add Thingy:91 X DFU command
This patch adds a new west command to flash the Thingy:91 X.
It automatically detects whether to update the nrf53 or the nrf91
and triggers the corresponding bootloader.
After the command is finished, the new application is running.
The command is used with a in the following way:
```
west thingy91x-dfu build/zephyr/dfu_application.zip
```

Signed-off-by: Maximilian Deubel <maximilian.deubel@nordicsemi.no>