applications: connectivity_bridge: Enable BLE on thingy91x
For enabling BLE on the thingy91x, the short range RF Front end module
needs to be enabled. This patch enables it when BLE is needed and
disables it otherwise. It does so only for the boards that have the
short range rf froend end module in device tree. The zephyr,user node is
used for this purpose.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>