tfm: Change the default TF-M profile type to be NOT_SET for 54L
The minimal profile type actually means consuming 32kB of FLASH, and
for 53/91 this is the minimal that TF-M can consume. But for 54L, with
a trustzone granularity of 4kB instead of 32kB, this is no longer the
minimal size that TF-M can consume.

Change the default TF-M profile type to be NOT_SET for 54L.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>