bluetooth: samples: add iso_rx disconnect cb.
Add callback that is triggered when an RX ISO channel
disconnects. The usecase is to start scanning for a new
connection. Scanning for a new connection is also done
in the ACL disconnection callback in this sample, but the
ISO channel might disconnect after the ACL, which will
prevent the central from starting to scan for a new ISO connection.

Signed-off-by: Henrik Lander <henrik.lander@nordicsemi.no>