lib: lte_link_control: Avoid unnecessary eDRX enable
lte_lc used to enable eDRX with AT+CEDRXS and send AT%XPTW immediately
after. This causes an unnecessary TAU when PTW is set same value that
network reported.
Now AT%XPTW is not sent if PTW value given by the network is the one
we are supposed to request.

Jira: NRF91-2041

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>