net: lwm2m_client_utils: Handle Block-wise FOTA on background
When we receive Block-Wise FOTA packets. We should immediately
respond to incomming data. If we instead start erasing modem FOTA
area, the server times out and starts retrying the first blockwise
packet. This looks like FOTA starts multiple times in a row as
the LwM2M engine does not detect duplicates on first block-wise
packet. It is handled as re-started Block-Wise.

Therefore handling of FOTA data is moved to system worker queue,
so the LwM2M engine have time to respond to first packet.
Semaphore might block the write request on the second packet, but
if that is retryed, LwM2M engine drops the duplicates.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>