nrf_security: Support single-threaded Cracen
Support Cracen when multithreading is not enabled.

Created two macros for mutex handling which are empty
when CONFIG_MULTITHREADING is disabled.

Add busy wait implementation for the Cracen interrupts since
they use the Zephyr events APIs. The Zephyr events APIs require
either multithreading or the Zephyr user mode which is not supported
at the moment.

Ref: NCSDK-27385

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>