lib: at_monitor: use dedicated workqueue
This patch adds a dedicated workqueue to dispatch AT notifications.

Earlier, AT notifications were dispatched from the system workqueue.
This was RAM-efficient but led to problems because of the
system workqueue "popularity" that made it used for many tasks,
including blocking operations. Blocking the system workqueue would
halt the dispatching of AT notifications, causing at_monitor to assert
after running out of heap.

Dispatching from a dedicated workqueue mitigates this problem,
which fundamentally lies in the application performing blocking
operations in the system workqueue, at the cost of some RAM.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>