Samples: Bluetooth: Mesh: add nRF54L15 support to light_ctrl
This adds support for the nRF54L15 PDK to the light_ctrl sample. It
changes the interrupt line used, as the one that was there was not
available on the nRF54L15.

Note: Currently, there is an issue with PWM LEDs on the nRF54L15 PDK
where the LED dims inverted. I have not handled this here, since the
same incorrect behavior shows up in other samples and I expect it to be
fixed in Zephyr.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>