lib: nrf_modem: trace: uart: let traces be processed in chunks
Trace data is processed in chunks.
Every time a whole chunk of trace data is processed, it is freed
from the shared memory. By reducing or tweaing the chunk size,
it is possible to free the shared memory more often, albeit by a
smaller amount. This, however, can improve the availability of
shared memory, thus reducing the chance of losing traces.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>