modules: memfault: Override MEMFAULT_HTTP_USES_MBEDTLS config default
Having mbed TLS enabled in the system does not automatically mean it's
used for HTTPS communication in memfault. If offloaded sockets are used,
it's not the case regardless of mbed TLS configurations.

Override the default value for MEMFAULT_HTTP_USES_MBEDTLS config in case
NRF_MODEM_LIB is enabled to n. This config simply enables some extra
checks for mbed TLS configuration (whether it fits memfault needs), but
since it's not really used for offloaded sockets, it can generate
false errors. By default, the config is set to y whenever mbed TLS is
enabled, which is oversimplified condition for our case.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>