tfm: 54l: Add Zephyr's soc.c to the TF-M build system
Add Zephyr's soc.c to the TF-M build system and TF-M image.

Zephyr's soc.c will do misc. power and clock configuration based on
the Kconfig and DT of the non-secure image.

But the non-secure image does not have access to NRF_OSCILLATORS or
NRF_REGULATORS for security reasons.

To solve this we have ported soc.c to TF-M.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>