tf-m: 54L: Enable ITS encryption by default
On old platforms using the HUK for encrypting the ITS requires extra
code, but on 54L15 it works with no further additions by a customer.
To follow a secure by default practice the ITS storage is encrypted by
default. If this is to much overhead then it call be disabled if wanted.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>