Samples: Bluetooth: Mesh: Add FWID print on DFU build
Adds printing of the Firmware ID to sysbuild system for the Mesh
DFU samples. The FWID is extracted from the .config file and
printed in the bytestring format required by the mesh shell DFU
interface. The intention is that this will make it easier for users
to configure and run mesh DFU.

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>