net: lib: nrf_cloud_coap: Enable SO_KEEPOPEN
Enable the SO_KEEPOPEN socket option to reduce the need
to close and reopen the socket and perform DTLS
negotiation whenever the PDN disconnects and reconnects or
LTE goes offline and online.

Store paused condition whether session pause/resume succeeded
or not. Use to determine if ready to perform a request.

Set a finite timeout for callback from NON transfers.

When CoAP socket is paused, prevent outgoing requests.

Fix error handling in client_transfer() so serial_sem
is not left open.

Cancel ongoing coap_client requests when pausing or disconnecting.

If we cannot save the CID when pausing, we will need
to do a full reconnect.

When receiving a callback, validate the user_data thoroughly.

Initialize the result code to an error value rather than
start uninitialized.

Jira: IRIS-8593

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>