drivers: wifi: Fix interrupt unregistration
During interrupt unregistration we need to remove the interrupt handler
first to avoid further interrupts and then wait for any pending items in
the workqueue to complete before returning to the caller.

Use "sync" variant for waiting till pending work is done and swap the
operations to remove the interrupt handler.

Also, check for the return value of handler unregistration.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>