applications: Matter Bridge: avoid runtime crash caused by bad alloc
When adding multiple devices to the bridge it may happen that
we run out of heap. In such case the allocation of mDataVersion
may fail and we end up with nullptr that is later passed to the
Span causing its constructor to abort().
Avoid abort() by adding a sanity check and more error handing
to the BridgeManager.

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>