applications: Fixed writing the bridged device node label
It is not possible to modify the bridged device node label e.g.
using controller commands, despite the spec says this is R/W
attribute.

Changed attribute type to writable and exteded implementation
to allow handling write requests for it. Additionally added
parsing the write requests with string type, as it turned out
to not work properly.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>