lib: lte_link_control: Fix null pointer access
The reg_status param is allowed to be set to NULL by the call. So we
should not derefence it without a NULL check. This patch uses the status
variable instead of dereferencing the reg_status parameter.

Added unit test to cover this scenario. The unit test fails with a
segmentation fault without this fix.

Found as violation of CWE-476 by SonarCloud.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>