samples: dtm: nrf5340 align to sysbuild SHIELD
The SHIELD variable is applied to all images when sysbuild is used.
The shield implementation provides appropriate overlays to each core
for given shield. This includes the pin forwarding for the application
core. Because of that the pin forwarding for nrf5340 in application
core is removed from the dtm application. The dtm application now
relies on providing the proper pin forwarding for nrf5340 by
the implementation of the shield.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>