shields: nrf2240ek: align to sysbuild
Zephyr checks if given shield is supported by searching
`<shield>.overlay` file. This file is the applied to device tree for
given core of the SoC. Without the sysbuild the SHIELD could be applied
differently to each core of the multicore-SoCs. With the sysbuild
the SHIELD variable is passed to all images involved in the build.

For multi-core SoCs like nRF5340 the approach where overlay is held
in <shield>.overlay file not hold because the content of overlay
to be applied differs between cores while the SHIELD variable is the
same for all cores.

The approach for the support of the shield is changed so that
appropriate overlays are applied on per-board basis.

The <shield>.overlay still exists, because Zephyr detects if shield is
supported by searching this file, but since now the file is empty.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>