net: lwm2m_client_utils: Handle retried FOTA writes
When Block N=0 is retransmitted, our FOTA handler might see that
as a completely new FOTA process, if previous one failed.
We should instead return failure if the FOTA is retried as
we already know that we failed the previous time.

If we instead accept the first block, then retries are just
skipped and handled normally.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>