Bluetooth: Mesh: Fix scene recall for LLC server
When recalling a scene during when the controller is not in standby, the
Light Lightness Server would turn off the light even if the scene had
the light turned on.

This commit fixes the issue by removing a check for whether the light is
on when recalling a scene. The check was added to prevent recalling the
same scene multiple times, but it is not necessary as this is checked in
in `bt_mesh_scene_srv_set` (in `scene_srv.c`).

Signed-off-by: Håvard Reierstad <haavard.reierstad@nordicsemi.no>