applications: serial_lte_modem: data mode support for object /19
The test specification of a certain carrier requires to test the upper
limit of data set through the AT#XCARRIER="app_data_set". This cannot
be set through a single command and so must be done via data mode.
Change the format of the command to accept the target URI path as
a string to be tokenized.

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>