bluetooth: Fix potential nullptr dereference in mds
MDS allows registration of a user supplied callback function to control
acces to the Memfault characteristic and descriptors data. However, when
the presence of this callback registration is tested in
data_export_write, data_export_ccc_write, and mds_read, it is falsely
assumed that the callback structure itself is always registered.
Since this can considered programming mistake, an assert would also
resolve this but that would bypass the optional nature of the
registration API.

Signed-off-by: Jelle De Vleeschouwer <jelle.devleeschouwer@classified-cycling.cc>
Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>