Bluetooth: Mesh: Don't start LC regulator on powerup
After PR #9865, the PI-regulator should only start once the server has
started receiving ambient lux level data, and not when the LC starts.
This was fixed for initial boot and when manually enabling the LC
server, however the Power OnOff handling was not updated to align with
this, causing every subsequent reboot to indiscriminately start the
regulator at boot without waiting for sensor data. This is fixed with
this PR.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>