samples: matter: Provide a fix for unacceptable booting time
Currently, without tickless kernel, the SYSCOUNTER value
after the software reset is not set properly and due to that
the first system interrupt is not called in the proper time,
the SYSCOUNTER value is set to the value from before
reset + 1. Hence, the reboot time increases more and more.

To avoid it enable tickles kernel for mcuboot.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>