bluetooth: fast_pair: separate subsequent pairing feature into Kconfig
Added a new Kconfig option for the Fast Pair Subsequent Pairing
feature. As a result, this previously supported feature is now
configurable and can be disabled for the locator tag use case,
which does not support subsequent pairing.

Aligned the Fast Pair Advertising Provider with the new Kconfig option
for the Fast Pair service, which controls the Subsequent Pairing
feature.

Removed the unavailable API call of the Fast Pair Advertising Provider
from the Fast Pair Locator Tag sample.

Ref: NCSDK-27955

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>