samples: matter: optimize the factory reset time
Currently the Matter core uses generic implementation
of GroupDataProvider that performs multiple operations
on persistent storage (including erasing of the related keys)
even though the KVS is erased at the end and the device is
rebooted. Implemented custom GroupDataProvider that can be
configured at runtime to avoid redundant and time-consuming tasks.

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>