mpsl: Measure HFCLK latency on boot
On boot, this measures the time needed for the HFXO to be ready for
radio activity. MPSL will then use this value to decide when to request
the HFCLK before an event.

Using mpsl_clock_hfclk_latency_set() directly
is ineffective for this purpose, because the user needs to know
about various factors that can affect the required latency/distance
such as HFXODEBOUNCE or HFXOCNT.

With our DKs this reduces the time spent with the HFCLK running by
1ms per event (7e-8 mAh per connection event/advertisement/etc)

Signed-off-by: Olivier Lesage <olivier.lesage@nordicsemi.no>