nrf_security: cracen: SRP role fixes
The set_password_key function is called by the
cracen_pake_setup function. The role of the operation
is set after the setup function, so this change uses
the size of the buffer to distinguish what input it
gets during the setup phase.

Also adds the calculation of v based on k if the
role is server to comply with the Oberon implementation.

Ref: NCSDK-26665

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>