nrf_security: cracen: SRP key management fixups
Update the import_srp_key function in Cracen to
better handle the key bits. In the PSA spec setting
the key bits is not mandatory. This change updates the
driver to allow getting a 0 bits key and it makes sure
that it sets the key bits correctly during the import.

Ref: NCSDK-26665

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>