editorconfig: Add editorconfig file
Adding an editor config file to improve developer experience.
Especially nice for editors that default to a tab-size of 4
instead of 8.

The file added in this commit is an exact copy of the one
found in zephyr.

EditorConfig (https://editorconfig.org) is a widely supported
configuration tool that helps to ensure better consistency amongst
developers by auto-configuring an editor to match the original/intended
style i.e. tab-width etc

Lots of editors support this natively and those that do not probably
have a plugin or extension.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>