mqtt_helper: Prevent mqtt_helper_thread hang on mqtt_live fail
If mqtt_live fails with an error other than -EAGAIN
(for example -ENOTCONN), the poll loop exits.

connection_poll_sem is still held, and no attempt
is made to abort the connection, so it is not possible
for the loop to recover.

Instead, we call mqtt_abort on any occasion that we
break out of the loop.

This means that a disconnect event occurs, allowing
us to gracefully recover.

Signed-off-by: Hayden Ball <hayden@playerdata.com>