Bluetooth: Mesh: add VERSION files to mesh samples
This commit adds VERSION files to the samples that were missing it, as
well as setting the prj config CONFIG_NCS_APPLICATION_BOOT_BANNER_STRING
accordingly.

Currently the boot banner is not printed when using CONFIG_BT_MESH_SHELL
or CONFIG_SHELL, which affects the chat and sensor_client samples. This
will be fixed at a later point.

Signed-off-by: Kyra Lengfeld <kyra.lengfeld@nordicsemi.no>