sysbuild: Split variant image logic
A variant image copies several properties of the original image. This
includes cache variables, which should be copied as late as possible,
because more of these variables can be set during sysbuild processing -
for example, through the original image's `sysbuild.cmake`, which may be
sourced after `ExternalNcsVariantProject_Add()` was already called.

The latest possible point in sysbuild would be a PRE_IMAGE_CMAKE hook
before the variant image is configured. Create this hook in NCS and move
the relevant logic there.

Ref: NCSDK-28495
Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>