nrf_security: Guard IKG function with key type
The IKG in Cracen only supports one asymmetric key
type, the SECP256R1. This guards the function call
to the IKG with the relevant configuration. This is
important because the PSA signing APIs can be used
with algorithms like EDDSA which have lower footprint
so it can reduce the flash usage when ECDSA is not
needed.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>