partition_manager: TrustZone HW model 1 workaround
The configurations: NRF_TRUSTZONE_FLASH_REGION_SIZE
and NRF_TRUSTZONE_RAM_REGION_SIZE are added after the
release and they are not included in the hardware
model 1. This is a workaround for this use case,
since we don't expect to support the nRF54L15 devices
with this model this workaround should work for the
older generation of devices.

Ref: NCSDK-28479

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>