mpsl: cx: restore dependency on nrf_radio_coex devicetree
In the previous change, I removed the dependency on the
"nrf_radio_coex" devicetree node from MPSL_CX Kconfig in
order to be able to use the software coexistence without any
DTS changes.

This, however, violates the least astonishment principle
because if someone enables one of the preexisting coex
implementations in Kconfig but forgets to update the DTS
or satisfy other dependencies, the build automatically
switches to the software coex (with only Kconfig warning).

It's probably more consistent to always assume that the
nrf_radio_coex must be defined whenever MPSL_CX is set.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>