snippets: Add snippet for nRF54H20 PSA rng
Add a snippet for enabling the PSA RNG as the
Zephyr chosen entropy for nRF54H20. This is meant
for users who use a Secure domain firmware version
with the SSF server enabled.

The snippet will be removed later and the logic
will be included as default in the dts/defconf files
for NRF54H20.

Ref: NCSDK-26837

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>