partition_manager: Move ns storage partitions for nRF54L15
When we have TF-M enabled with the nRF54L15 we want to minimize
how many secure/non-secure regions we use, so in TF-M we make
the assumption that there is one secure and one non-secure region.

This updates the non-secure storage partitions (settings/nvs) in
nRF54L15 so that they are placed after the non-secure application.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>