lib: location: Fix GNSS obstructed visibility detection
Fixed a bug which caused only part of the satellites to be iterated
when counting the number of tracked satellites. This was caused by
a problem in the satellite C/N0 level check. The check has now been
removed, because it is no longer relevant.

Fixed a bug which caused visiblity detection to be sometimes performed
twice.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>