doc: Make all the tables in SML page manual
We have had issues with incorrect SML table values
due to the update being through nightly build,
especially in release branches. So  until the script is
optimised, the only way to fix this issue is to make
the tables manual (grid tables). This PR converts the present values
into manual table values.

Signed-off-by: Uma Praseeda <uma.praseeda@nordicsemi.no>