samples: bluetooth: peripheral_mds: Set device ID at build time
The device ID when using the MAC address was Unknown after the system
booted because the device has not been given a MAC address yet at the
time the device ID is set (system init). This change removes the option
for using a MAC address as the device id for BLE devices, and sets a
static device id for the BLE sample app.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>