boards: thingy91x: remove external_flash static partition
When using the start-to-end placement strategy in the external_flash
region, a partition named external_flash will be automatically created,
filling the remaining address space of the external_flash region.

The external_flash partition we had defined in the Thingy:91's static
partitions covered the entire external flash space. This caused dynamic
partitions, such as modem_trace, to be placed outside the external flash
address space instead of following the last static partition.

With this fix, a dynamic partition in the external_flash region will be
placed directly after the last statically defined partition. Then, a
partition named external_flash will be created to fill the remaining
space.

CIA-1330

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>