Bluetooth: radio_notification_callback: Refactor to not use EGUs
By using the the connection anchor point events instead of the
connection connection event trigger functionality the implementation
can be refactored to not use an EGU instance.
This approach will also scale better to multicore architectures.

This approach also uses less FLASH.
For the radio_notification_cb sample on 53, the FLASH usage
is reduced with almost 400 bytes.

For multicore devices that have different time domains on different
cores, a translation mechanism is needed. This is currently not present.
That is, we don't support this feauture on 53 app core, but it works
nicely on 54 as that uses the global GRTC.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>