lib: lte_link_control: eDRX request not in workqueue in power off
lte_lc_edrx_req() will be run directly in lte_lc_edrx_on_modem_cfun()
when power off is issued. Earlier it was submitted to a work queue
and run in lte_lc_work_q.
Work queue was originally added because we wanted to avoid sending
AT commands in the callback. However, when modem is powered off,
we are not expecting AT notifications that could cause an assertion or
missing notification.
This was added in PR #15203

Jira: NCSIDB-1347

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>