boards: thingy91x: use nrfutil runner
This patch uses the nrfutil runner by default for the Thingy:91x board.
Also, the device family option was removed for the nrfjprog runner.
The device family option is dangerous to use and should be avoided.
For nrfutil-device, a safeguard is being put in place to remedy that
problem, but nrfjprog will not be updated in that way.

Signed-off-by: Maximilian Deubel <maximilian.deubel@nordicsemi.no>