samples: bluetooth: radio_coex_1wire: Use EGU instead of TIMER
This change uses egu interrupts instead of timer counter to count radio
ready events.

This is a preparation commit for nRF54L15. In nRF54L15 the timer on the
same PD is occupied by SDC. To access the nearest available timer we
need to use PPIB.

Also, since MPSL uses predefined channels for radio events on hardware
with DPPIC support, the predefined channel needs to be used with
nrfx_gppi library. But the library doesn't support predefined channels
at this moment, it always allocates a new channel, therefore can't be
used at the moment. Because of this, we have to hardcode peripherals
(DPPIC20, PPIB11, PPIB21 and TIMER20) in the sample which doesn't look
correct. Also, bypassing nrfx_gppi may create conflicts if someone else
is gets allocated the same channel through nrfx_gppi (e.g. if an
implementor copy-pastes this code and uses it together with nrfx_gppi
library). With EGU we don't have these issues on nRF54L15.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>