Bluetooth: Controller: Add device tree dependency to selection
This commit makes the device tree configuration decide
which link layer is compiled in. We do this to avoid
hard-to-understand linker errors referencing device tree nodes.

The following configuration creates a difficult to parse
linker error:
* BT_LL_CHOICE contains multiple entries, BT_LL_SOFTDEVICE
  is selected.
* Each BT_LL_CHOICE has its own unique device tree node with its
  own "compatible".
* Only one of the link layer device tree nodes has status "okay",
  but not the one corresponding to BT_LL_SOFTDEVICE.

The linker error indicates that code using the HCI driver fails
to link with the controller. This because the HCI driver device
tree node references the link layer selected in devicetree which
is not compiled in.

By adding a dependendency to the device tree node, this can no longer
happen. Instead, if BT_LL_SOFTDEVICE is now selected in a configuration
file, a Kconfig warning will be issued:

```
warning: The choice symbol BT_LL_SOFTDEVICE
(defined at
/home/ruge/ncs/nrf/subsys/bluetooth/controller/Kconfig:11) was
selected (set =y), but BT_LL_SW_SPLIT (defined at
subsys/bluetooth/controller/Kconfig:129)
ended up as the choice selection.
```

This should be easier to understand than:

```
...(hci_core.c.obj):(.data.bt_dev+0x16c):
undefined reference to `__device_dts_ord_132'
```

A similar change has been made for the Zephyr Link Layer in
https://github.com/zephyrproject-rtos/zephyr/pull/78897.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>