nrf_security: Add SSF client without nrf_security
Add configuration to allow enabling the SSF PSA client
when nrf_security is not enabled.
This is particularly useful for the applications that only
want to use the PSA rng and no other crypto. Enabling
nrf_security in these applications will result to an
increased application footprint and configuration complexity
without any reason.

The approach that I took here is to avoid doing the conversion
of Kconfig->Cmake->Header that we do in nrf_security.
First the PSA SSF client is not configurable at all so
adding configuration options doesn't give anything.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>