samples: crypto: psa_tls: Remove Oberon PSA case
Remove the test case for Oberon PSA on nRF54L15
since it is not fully suppoted. In nRF54L15
the random driver for Cracen has dependencies on
other functionality of the Cracen driver which
means that you are forced to enable more driver
features in order to get random. Enabling features
in Cracen automatically disable the equivalent features
in Oberon.
Since this case requires random from Cracen most of the
Oberon functionality of Oberon will be disabled anyway
so we disable it here.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>