samples: cellular: nrf_cloud_multi_service: reduce prov rate
Once device has connected to the cloud, slow down the
provisioning check.

Improve logging of shadow and FOTA polling intervals
for CoAP.

Make CoAP shadow and FOTA check intervals the same as
the sensor update interval.

Improve handling of credentials check to check for new error
code for bad CA size, and also turn off network if credentials
are bad and the nrf_provisioning service is not used.

Jira: IRIS-9151

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>