lib: sms: Possibility to disable status reports
Clients can decide if they want to receive SMS-STATUS-REPORT
with CONFIG_SMS_STATUS_REPORT.
When this is not set, SMS registration is done with AT+CNMI=3,2,0,0
so delivery reports are not sent to this library,
and SMS header will not include a bit to request SMS delivery reports.
By default, this is ON so the change is backwards compatible.

Jira: NCSDK-29508

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>