tfm: Fix nrf54L15 linking with nrf_security targets
This fixes an issue for the nRF54L15 when it is build
with TFM and the hardware unique key library is also
used. The hardware unique key library uses PSA headers
which means that we need to make sure that the nrf_security
generated headers are being included before the other PSA
related headers.

This links the nrf_security target before it links the TFM
platform targets.

Ref: NCSDK-29514

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>