lib: sms: ACK before APP callback
Received message and status reports have been first scheduled to queue
before the acknowledgement of the message with AT+CNMA=1.
The order is changed because acknowledgement to network is more
important and can be delayed if application does too many things in
the callback, which it shouldn't do anyway.

Jira: NCSDK-29532

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>