suit: Return volatile key IDs from MCI
In some cases, there is a need to use volatile keys for signing
manifests. In such case, it is the PSA API that sets the key ID
value, not the manifest.
Add a simple MCI API helper function to return the assigned key ID
and handlers inside manifest authentication routines.

Ref: NCSDK-NONE

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>