nrf_security: Refactor builtin key handling
This updates the way we handle the configuration
MBEDTLS_PSA_CRYPTO_BUILTIN_KEYS so that it follows the
same logic as all the other PSA Crypto core configurations.
Also does minor cleanups for this option.

It also renames the configuration option
MBEDTLS_ENABLE_BUILTIN_KEYS->MBEDTLS_PSA_CRYPTO_BUILTIN_KEYS.

This is a promptless option so no need to update any
sample/application because of this.

Ref: NCSDK-29543

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>

Update subsys/nrf_security/Kconfig

Co-authored-by: Tomi Fontanilles <tomi.f@tutanota.com>