samples: wifi: provisioning: softap: increase TF-M flash partition size
This fixes twister that was failing due to an overflow.
This appeared with the update of TF-M to 2.1.1.

A slight increase in code size made it grow past the next 16KiB block,
resulting in a corresponding increase in S image flash usage.
The granularity of allocated size on 53/91 Series devices is 16KiB.
As such, the TF-M flash partition size is increased by that amount.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>