scripts: ncs_commands: Exit in certain conditions
The https://github.com/nrfconnect/action-oss-history action uses "west
ncs-loot" to check the history of OSS repos. The action needs to fail
whenever something is amiss in the run, so bump the severity of certain
(non-recoverable) errors and a warning in order to ensure we never
ignore errors in the action.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>