net: wifi_credentials: Fix commands mixup
With the recent changes to Wi-Fi shell to use the separate command
declaration so that Wi-Fi commands can be extended externally, this
causes some sort of conflict with wifi_cred, possibly because both
command parent names start with "wifi", few commands from "wifi" end up
wth "wifi_cred".

As we debug the root cause, temporarily renamed the parent to
"nwifi_cred" to avoid this mixup, the shell command is still
"wifi_cred", so, this is acceptable.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>