suit: root manifest template support for independent updates
This commit modifies SUIT the root manifest template to support
the scenario of updating the radio and application manifests
one at a time. It assumes that in such case the radio image
is always updated before the application image.

It also adds several version checks for the candidate image - these
are examplary checks, it is suggested to change them to be tailored
to the user needs:
- Ensuring that a compatible Nordic soc binaries bundle is installed
  (currently version 0.6.5 of the Nordic Top manifest)
- If the candidate only contains the radio core image, ensure that
  its major version of the candidate radio manifest is at most greater
  by 1 from the installed application manifest major version.
- If the candidate only contains the application core image, ensure that
  its semantic version of the candidate application manifest exactly
  matches the installed radio manifest semantic version.

To separate the application/radio manifests from the root.suit
envelope the newly added suit-generator payload_extract command
can be used.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>