Bluetooth: GATT DM: Schedule host cbs in work q
Moves host calls in discovery manager callbacks to be scheduled in
either a GATT DM-specific work queue or the system work queue depending
on a new Kconfig option `BT_GATT_DM_WORKQ_CHOICE`. This is done to
improve robustness and avoid possible concurrency issues.

Using the system work queue will not increase the memory footprint, but
is marked as experimental as the callbacks can potentially block, which
is not recommended in the syswq.

Signed-off-by: Håvard Reierstad <haavard.reierstad@nordicsemi.no>