appplications: sdp: gpio: add asm_check to flpr_egpio
Add asm_check target as a depency for flpr_egpio application.
This will generate new ASM file for every source file in
hrt folder and check if the new ones are the same as the
one saved in git. For every file that differs a warning
is printed.

By making asm_check a dependency of flpr_egpio application
asm_check will be run both for sysbuild application including
eGPIO and for stand-alone SDP application before the main
build starts.

Signed-off-by: Magdalena Pastula <magdalena.pastula@nordicsemi.no>