samples: mpsl: timeslot: Fix build errors for 54L and 54H SoCs
The sample application didn't build for both 54H20 and 54L15 SoCs.
There are different timer instances used by MPSL for new SoCs.
Also the logs ofload ISR in 54H/54L are done by different peripheral.

The commit fixes the build issues.

By the way the commit enables the sample to be run in CI and adds
missing information to the readme file.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>