samples: matter: Add guide on how to flash factory data on nRF54H
Until we do not have a proper solution for merging the factory data
hex file with the firmware hex file, we need to flash the factory
data manually. This commit allows doing that on nRF54H20 DK and
describes how to do it in Matter samples.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>
(cherry picked from commit 5736ed3)