Bluetooth: Mesh: Fix recall in light_ctrl_srv
Previously, the scene recall was not working as expected in the
following case: If a single scene is stored when the light is on, and
then the light is dimmed/turned off by the controller, the current
scene would not be invalidated and recalling the scene would do nothing.

This commit fixes the issue by invalidating the current scene when the
light controller is re-enabled as a result of the resume timeout.

Signed-off-by: Håvard Reierstad <haavard.reierstad@nordicsemi.no>
(cherry picked from commit a07804cccb31616bff75ffb0fcfb81febdbdc016)