suit: Recovery run from APP_LOCAL_3 manifest
This commit changes the way the recovery application
image is connected to the APP_RECOVERY manifest.

In the new approach, it is not accessed directly from the
APP_RECOVERY manifest - instead, the APP_RECOVERY manifest
is connected to the APP_LOCAL_3 manifest, which then orchestrates
invoking and installing of the application recovery image.

It is still possible to use the old approach by setting
SB_CONFIG_SUIT_RECOVERY_APPLICATION_IMAGE_MANIFEST_APP_RECOVERY
during build.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>