suit: Allow for local manifests installation from recovery
This commit allows the APP_RECOVERY manifest to orchestrate
the installation of the APP_LOCAL/RAD_LOCAL manifests.

This is needed for cases where there is a "companion image"
in the system, which can be invoked by both the root and recovery
manifests. In this case, such images should only be installed
from recovery mode.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>