samples/applications: Optimize memory for nRF91+nRF70 positioning
nRF91 with nRF70 shield was initially made to work for positioning
purposes and a lot of memory has been reserved for the operations.
nRF70 driver has been optimized a lot since then so we can reduce many
of the stack sizes as well as heap for these scan only configurations.

The saving is ~25kB of RAM. The main contributors are:
* 0-15kB: Lowering the kernel heap (CONFIG_HEAP_MEM_POOL_SIZE).
          In most cases size was reduced by 10kB.
* 4kB: CONFIG_LOCATION_WORKQUEUE_STACK_SIZE from 8192 to 4096,
       which has been default without nRF70 shield
* 7kB: CONFIG_NET_TX_STACK_SIZE and CONFIG_NET_RX_STACK_SIZE
       from 4096 to 512
* ~1-2kB: Lowering CONFIG_NET_BUF_RX_COUNT, CONFIG_NET_BUF_TX_COUNT and
          CONFIG_NET_MAX_CONTEXTS

Jira: NCSDK-30166
Jira: NCSDK-30086
Jira: NCSDK-30171

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>