shields: nrf2240ek: no more explicit pin-forwarder shield
The nrf2240ek_fwd is no more necessary and deprecated. Build system
picks appropriate overlay (being either a pin-forwarder or the main
fem overlay) based on content of `boards` directory.
User does not need to pass different values for `-DSHIELD` for every
core on the SoC. It is handled on the shield level and passing just
`-DSHIELD=nrf2220ek` for all cores is enough.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>