actions: manifest: Align with upstream Zephyr
The manifest action now has many more features at version 1.6.0. Update
to the latest and start using a tree checkout to diff the manifest,
which is compatible with split manifests.

This is a second attempt after bf6df10f571b5a15110f0ac8555ba8b83de299c8
had to be reverted in f4744e2ae81cdcdc01071b6a05eb02905ac75776 due to
the fact that the actual action did not support selecting which projects
to import. This has now been fixed in:
https://github.com/zephyrproject-rtos/action-manifest/commit/1807b4c7d4d23b36c68587aa9018d3e74518f9ff
and included in v1.6.0 of the action.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>