samples: bluetooth: fast_pair: input_device: Improve nRF54H20 memory map
Puts fast pair provisioning partition right before the storage partition
and sets its size to 4 KiB to avoid gaps. Align the storage partition to
a 4KiB boundary as this is the sector size. This alignment is needed due
to a bug in the storage subsystem. Fixes the cpuapp_rw_partitions node
deletion (earlier it was done incorrectly).

Jira: NCSDK-30269

Signed-off-by: Aleksander Strzebonski <aleksander.strzebonski@nordicsemi.no>