samples: suit: smp_transfer: Use PRNG entropy
This sample require entropy from Zephyr, in nRF54h20
this is provided by PSA RNG driver and from the secure domain.

The PSA RNG driver brings IPC dependencies which increase the
flash footprint of this sample and this was not an acceptable
increase for the mainttainers of the sample.

It was concluded that as a temporary solution this sample  will keep using
the non cryptographically secure, deterministic software RNG.

The dependency on the PRNG node needs to be removed later and it tracked
in NCSDK-30805

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>