samples: Bluetooth: update scanning_while_connecting
Move flag that protects application from establishing
a new connection when there is already a connection
establishment process pending. This is moved to after
the if to check that we have established all connections.
Otherwise, we get in the scan_recv call back an attempt
to establish a connection when we have used all
connection objects before we stop scanning in main.

Signed-off-by: Henrik Lander <henrik.lander@nordicsemi.no>