bluetooth: host_extension: RTCx time diff may be negative
Radio notification callback host extension checks the offset between
system clock and second RTC instance. The order of RTCx startup depends
on MPSL being used as clock control driver. If the MPSL is the driver
it is started at PRE_KERNEL_1 init level that makes RTC0 instance
to start first. In case of use of nrf clock control the MPSL is
initialized at POST_KERNEL init level that makes the RTC1 to be started
first.

To make the timer correction to be right, change unsigned types of
variables used to store the difference to be integer. Then we don't
depend on the RTC initialization order.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>