samples: suit: smp_transfer: Change hci_ipc stack defaults
Enabling real entropy for the radio core through the ssf_client and
the secure domain increased the stack requirements of the hci_ipc
used in this sample.

I couldn't run THREAD_ANALYZER in this application because of flash
overflows and other issues. I did practical tests with 50 byte intervals
and I know that 900 bytes is the least memory that could boot the radio
core.

I updated this to have the same configuration as the ipc_radio (2048
bytes)
application since the usage of the hci_ipc here will be replaced
with the ipc_radio later.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>