mpsl: clock_ctrl: Use new nrf2 clock ctrl HFXO zli api
There were no API in nrf2 clock control that allows to requrest
the HFXO from zero latency interrupts. That was a blocker for MPSL.
Since the API was added, change the implementation of clock control
integration layer to use it.

Mind that the new API is similar to nRF52/nRF53 APIs:
z_nrf_clock_bt_ctlr_hf_request and z_nrf_clock_bt_ctlr_hf_release.
It does not provide feedback if the HFXO ramp-up is done.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>