applications: nrf5340_audio: Fix audio sync timer off by one tick
Fix audio sync timer off by one tick when TASKS_CAPTURE and
TASKS_CLEAR are in a race condition.

TASKS_CAPTURE is prioritized hence can have +30 us more when
RTC tick to trigger TASKS_CLEAR overlap with the capture of
the value.

The fix here initiates the RTC and TIMER sync only when a
capture is requested by the audio subsystem or the timestamp
request. This is in contrast to earlier implementation when
TIMER was sync on every RTC tick.

If a capture is requested in short duration of 30 us then
the race will still be present, but a subsequent fix for it
will be provided in a new commit.

Signed-off-by: Vinayak Kariappa Chettimada <vinayak.kariappa.chettimada@nordicsemi.no>