mpsl: Support mpsl_pin_debug on nRF54L DK
This commit adds support for using pin debug on nRF54L.
The code has been extracted out to a separate file because has a
couple of things that are different compared to nRF52 and nRF53:
- It needs to setup a PPI bridge between two domains.
  This is done through `nrfx_gppi_edge_connection_setup()`
- It needs to allocate DPPI channels in the domain where the
  GPIO pin is located.

The code hardcodes to use PORT1. We should consider moving
the pin allocations to device tree to make this more flexible.

The code should be reusable for nRF54H as well.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>