drivers: serial: uart_nrf_sw_lpuart: Fix early TX abort
Fix case when uart_tx_abort is called before actual transfer
is started. In that case, internal call to uart_tx_abort was
returning -EFAULT and this error was returned to the user.
This error is now overwritten to 0 since that is a correct
case when TX is aborted during transfer setup.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>