net: lib: nrf_cloud: gateway improvements
Remove most customizations for the nRF Cloud gateway
because they are no longer necessary. The exception is
the proper suffix on the cloud to device topic.

Fix bug in JSONata transform handling during connection
startup -- only the MQTT control channel needs to be
connected, not also the data channel.

Jira: IRIS-9993

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>