nrf_security: Refactor get_opaque_size function
This refactors the function cracen_get_opaque_size so that it
returns psa_status_t and not a size_t value. With ths previous
implementation it was not possible to distinguish between
a key with invalid arguments or a key which is revoked.
Revocation support will be added soon and this separation is
needed.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>