emds: Fix size calculation required by emds entries on RRAM
The total size required by emds entries calculated more than what is
actually being used. For RRAM, we don't need to align the data with
write block size. We can write as many bytes as we want up to the size
of write block size, thus we don't need to align writes with write
block size.

Signed-off-by: alperen sener <alperen.sener@nordicsemi.no>