lib: modem_key_mgmt: fix duplicate deletion of last cred type
Because of the ok\r\n in response from modem, the while loop iterates
an extra round with scanf not matching any arguments. As the sectag
matched the last credential is attempted deleted again and the function
return an error. This commit fixes the issue by only deleting the
credential in the modem if we match the sectag and credential type in
scanf.

Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>