sdfw_services: psa_crypto: use new tags for pointer-to-const members
This lets the server loosen the required permissions when validating
pointers-to-const parameters so that only read permissions are required
for these.

Note that this CDDL update is not backwards compatible, and the
service version is bumped to 2.

Ref: NCSDK-31182

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>
(cherry picked from commit 48253248ecc493e0c3fed104f8ff090d5a55964d)