lib: nrf_modem_lib: Use debug logs of expected errnos on backend write
We use -ENOSPC and -ENOSPC to signal from the backend that there
is no space in backend or it is not properly connected. This can
occur from time to time and is not always an error. Though, it can
be noisy in log output. This commit replaces the error logs with
debug level and adds a hint that the backend should be properly
connected.

Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>