samples: wifi: Disable PSA RNG for all samples
The PSA RNG will be the default entropy provider for
nrf54h20. This change affects the crypto functionality
in general since it makes the secure domain the sole provider
of crypto through the PSA APIs. This has the sideffect that the
legacy mbecrypto APIs are not currently supported.

Since wifi relies on the mbedTLS legacy crypto functionality
we need to make sure that it continues to work as before.

In this change we disable the PSA RNG as the entropy provider and
we explicitely set it to be the PRNG. This makes sure that the crypto
funcionality is provided by the software implementation which supports
the legacy APIs and not from the secure domain.

This should not change anything functionaly. It only makes sure that the
wifi samples do not inherit enabling the PSA RNG as the entropy provider
by default.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>