bluetooth: controller: Add dependencies to Kconfig symbols
Add dependencies to SoftDevice Controller specific config entries so
that symbols that are not relevant for a given application
cannot be configured. This also removes these from the generated
configuration file.

This requires some change to source code
as well because some Kconfig entries are now undefined for
some application configurations.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>