nrf_security: Add PSA compatibility layer for SSF
There are two functions which are defined in the psa_crypto_core.h
and are implemented in psa_crypto.c which are used by the TLS
library.
These functions are:
psa_can_do_hash
psa_can_do_cipher

These functions just check if the drivers are initialized
before the relevant PSA crypto functions can be used.
In the case of SSF there is no initialization needed because
the PSA initialization happens inside the secure domain firmware
before the application boots.

These functions are added in a separate file since they only
exist to maintain compatibility with the PSA core from Oberon/mbedTLS
and they have don't need to forward any call to the secure domain.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>