doc: fem: add supported FEM table
The support for FEMs does not grow equally for all FEMs and SoC
families. To let customers know which FEMs are supported and on
which SoC Series a table under chapter
`Developing with Front-End Modules` is added to express this.
This table replaces some random sentences about SoCs and skyworks as
well.

The nRF2220 is mentioned there as well.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>