mpsl: pm: Change init API and adds uninit API MPSL PM
Refactor initialization API and add uninitialization.

Change the initialization API to do not allow call it
when it is already initialized or if uninitialization
is started but not completed. The uninitialization
API will wait for scheduled work queue uninit item
to complete. If that is late the API will timeout.

Uninitialize API will fail when the MPSL PM is not
in initialized state.

Unit tests are changed to cover changes in the API.

Signed-off-by: Knut Eldhuset <knut.eldhuset@nordicsemi.no>
Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>
Co-authored-by: Piotr Pryga <piotr.pryga@nordicsemi.no>