mpsl: clock_ctrl: Allow LFCLK timeout on nRF54H20 due to sysctrl IPC issue
Add temporary workaround that allows the LFCLK to timeout on nRF54H20
(if CLOCK_CONTROL_NRF2 is enabled). The potential timeout is not
an issue for now because the integration layer request the lowest
accuracy of LFCLK and such or better LFCLK should be runing from
boot of the radio core.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>