applications: nrf5340_audio: LED module for kits with other LED setup
The LED module is currently written in a way
that’s highly dependent on all of the LEDs present
on the audio DK, which prevents the application from
running on the nRF5340 DK. The module must be refactored
so that it can still show the state of the application with
a different selection of LEDs.

Currently the button handler expects kits to have five
buttons - which is true for the Audio DK, but not for the
nRF5340 DK. To have the audio application running on the
nRF5340 DK (and other kits in the future), the handler must
be refactored to not fail compilation when fewer buttons
are present, and reduce the application's overall dependency
on button availability.

Signed-off-by: Graham Wacey <graham.wacey@nordicsemi.no>