quarantine_zephyr: Add drivers.wifi.build.spi on native_sim/native
This test is not really useful in NCS context (checking if the nRF
Wi-Fi driver can be built successfully is done by other tests and
samples), and it requires the hal_atmel module to be pulled in
(because the WINC1500 driver requires two header files that are
located in that module). And the module weights ~87 MB, so it's
not reasonable to pull it in just for this one test.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>