nrf_security: Enable NRF_RNG_ENTROPY_DRIVER for fake entropy
The NRF_RNG_ENTROPY_DRIVER provides entropy though the PSA APIs.

This is used to seed/reseed the Oberon PSA random drivers
(CTR_DRBG/HMAC_DRBG). Before this was only used to allow enabling
this PSA driver with the devices which use the DT label 'rng'.

We extend that now to allow this driver to be used with the fake
entropy node which uses the DT label 'prng'.

This is needed to unblock development until real RNG is enabled
for all the needed applications/samples.

In practice this means that when 'prng' is enabled a device can
generate random numbers with the psa_generate_random call.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Signed-off-by: Frank Audun Kvamtrø <frank.kvamtro@nordicsemi.no>