Bluetooth: Mesh: Add FWID to DFU metadata json
This adds a mechanism for inserting a Firmware ID into the Mesh metadata
JSON file in the DFU zip, currently supporting a raw hex string set in
Kconfig, or an ID computed from the MCUboot version, aligned with what
the DFU samples use.

The samples are updated to use this new mechanism, and the
sample-specific script which printed a FWID is removed since it is no
longer needed (it has been superseeded by the new mechanism).

This also updates the DFU samples to insert the BT_COMPANY_ID in the
Firmware ID.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>