bootloader: bl_validation: Use correct S0/S1 symbols
The current implementation fails when building with
CONFIG_SECURE_BOOT=y and SB_CONFIG_SECURE_BOOT_APPCORE=y
on the 9160DK because the s0 region includes EMPTY_0 but
the s1 region has no equivalent.

This doesn't matter anyway, because the important thing
is the size of the s0_image and s1_image areas, not the
outer s0 and s1 areas.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>