samples: nrf_rpc: ps_client: fix BLE snippet after upmerge
The upmerge PR enabled BT_HCI for the protocols
serialization client sample, but the point of this sample
is to use BT_RPC_STACK, which enables RPC serialization
of Bluetooth host APIs, instead of the entire HCI-based
stack.

Restore the previous behavior and add necessary workarounds
to enable Bluetooth shell commands despite of not using
the HCI-based host implementation.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>