scripts: quarantine_integration: add nRF54L combos to the FP Locator Tag
Added the nRF54L10 and nRF54L15 board targets to the Fast Pair Locator
Tag sample quarantine in the release configuration variant. The
nRF54L05 board target should be sufficient to test in the integration
scope and prevent potential memory overflows as the smallest target.

Added the nRF54L05 and nRF54L15 board targets to the Fast Pair Locator
Tag sample quarantine in the debug configuration variant. The nRF54L10
board target should be sufficient to test in the integration
scope and prevent potential memory overflows as the target with the
smallest application partition in the debug configuration.

Ref: NCSDK-30305

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>