mpsl: clock_ctrl: Add MRAM high frequency request to MPSL
In the nRF54h20 SoC the non-volatile memory is located in a global
domain which by default configured by nrf2 clock control to run
with 64MHz. That makes a code execution time longer so it doesn't
fit into radio protocols time requirements. That causes e.g.
Bluetooth stack to assert in radio events.

To make sure the clock needed for MRAM (non-volatile memory in
nRF54h20 SoC) is always setup to highest avaialble frequency
MPSL will put a request for the clock to clock control driver.
In case there is any other user of the clock that runs from
radio core and allows to lower the global domain clock frequency,
the clock will still be stay as requested by MPSL.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>