hids: Check for slab buffer overflow
It's pretty easy to forget to specify all of the report types in
BT_HIDS_DEF() and overflowing those slab blocks can lead to corruption
in other slab pools leading to very hard to catch errors. Let's add a
number of safeguard checks (gated by CONFIG_ASSERT) to make sure this
type of overflow is detected.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>