dts: bindings: Update nordic,nrf-mpc
Update the MPC binding to cover both nRF54H20 and nRF7120 use cases.
Note that the MPC IP revisions used by the two devices are compatible.

Property `protection-region-size` is renamed to `override-granularity`,
a more appropriate term which comes from the datasheet. The description
is also corrected, as MPC stands for Memory Privilege Controller.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>