ci: doc: post preview comment after build finishes
As of today, preview comment was published just after PR is posted,
so if e.g. build fails, the comment may be misleading. Update the CI
jobs to perform this in 2 steps, so that the docpublish workflow takes
care of posting the comment.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
(cherry picked from commit ee6552c422c3af2b3d215ae585044b220166cda9)