nrf_security: drivers: cracen: Add workaround for ikg issues on 54l09/54l20
Add workaround to handle freezing issue on CRACEN Lite  devices.
Add compile time error for building with CRACEN Lite and not 54l09/54l20
As the developers should make a decisions if the workaround is needed.

This issue stems from the PKE-IKG interrupt not being cleared.
To handle this sx_pk_wait is changed to use polling and
the PKE-IKG interrupt is disabled except for when in PK mode.
In PK mode the interrupt can be cleared by hardware, which means it
can be used there.

Ref DLT-4048

Signed-off-by: Dag Erik Gjørvad <dag.erik.gjorvad@nordicsemi.no>