quarantine_zephyr: quarantine all secure_storage tests on 54L15 cpuflpr
With the latest changes the remaining test scenario that used to not be
filtered out is now overflowing.
This is not something we want to support anyway so quarantine all
secure_storage tests on that board target.

Also, replace the other quarantined secure_storage scenarios by the same
regexp to make sure to encompass them all.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>