Bluetooth: Samples: Add Advertising Coding Selection to Coded HR samples
Now that Advertising Coding Selection feature is supported. It
would be nice to have a sample that demonstrates how it can be
used. On the peripheral side this feature is used to indicate
coding scheme requirement of the advertiser on LE Coded PHY.
On the central side this feature allows the application
to obtain more detailed information of the advertisers primary
and secondary PHYs in the scan callback on LE Coded PHY.

Signed-off-by: Thomas Deppe <thomas.deppe@nordicsemi.no>