bluetooth: services: prefer ind over notif in RAS when relevant
This is based on the following text from RAS_v1.0.pdf

3.5.2 Ranging Data Overwritten behavior

If a RAS Server supports simultaneous enabling of indications and
notifications of the Ranging Data Overwritten characteristic, and the
client enabled both indications and notifications, then the RAS Server
shall use indications to report a Ranging Data Overwritten event.

3.4.2 Ranging Data Ready behavior

If the RAS Server supports simultaneous enabling of indications and
notifications of the Ranging Data Ready characteristic, and the client
enabled both indications and notifications, then the RAS Server shall
use indications to report a Ranging Data Ready event.

Signed-off-by: Olivier Lesage <olivier.lesage@nordicsemi.no>