mpsl: pm: Fix MPSL Pl uninit -NRF_ETIMEDOUT
There was a lock if a MPSL PM was uninitialized by
mpsl_lib_uninit() from hci_driver.c hci_driver_close().
The hci_driver_close() locks multi-threading by call
MULTITHREADING_LOCK_ACQUIRE(). The same mutex is acquired
by MPSL mpsl_low_prio_work_handler. The MPSL PM used
an uninit work item that was scheduled to MPSL low prio work
queue. The work was never executed during the MPSL PM uninit,
even though the uninit was waiting on a semaphore.
That was a lock that caused -NRF_ETIMEDOUT error to be
returned.

The mpsl_pm_uninit function can't be called without earlier
MULTITHREADING_LOCK_ACQUIRE() call. That is a prerequisite
to be fulfilled by a user.
On the other hand, the use of the lock allows to stop
using a work item during uninitialization of the MPSL PM.
The is not needed because if there is any work item in
the queue, then it will be handled after uninit is done.
It will have no effect on PM subsystem. These items will
be just removed from the MPSL work queue.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>