nrf_security: cracen: kmu: Fix ecdsa public key import
The ecdsa public key is 65 bytes long, but the kmu
only supported keys up to 32 bytes.
The first byte is Compact Binary Representation
where only one option is supported so this is removed
before storing and added back in when retrieved
This means we only use 4 slots instead of 5 to store the key

Fix bug in psa_crypto_driver_wrappers which meant kmu keys
could not be used for sign and verify hash.

Add support for storting  ecdsa keys using sha256 to the kmu

Signed-off-by: Dag Erik Gjørvad <dag.erik.gjorvad@nordicsemi.no>