bluetooth: services: Fix RAS ring buffer alloc overflow issue
When the ranging counter wrapped, alloc() would fail due to not
finding that the buffer with ranging counter 65535 could be freed.

To ensure the buffer selected to be freed is always the oldest, a new
approach is used. It's based on how many counts have elapsed relative
to the counter we are trying to allocate.

Signed-off-by: Olivier Lesage <olivier.lesage@nordicsemi.no>