samples: bluetooth: fast_pair: locator_tag: HW crypto MCUboot for nRF54L
Updated configurations for the nRF54L build targets in the Fast Pair
Locator Tag sample to use HW crypto in the MCUboot bootloader.

Aligned the partition layout of the affected targets to fit the MCUboot
image into its partition and reserve the appropriate room for the
eventual bootloader image growth in future NCS releases.

This change breaks the backwards compatibility as it changes the
MCUboot signature type from RSA to ED22519 and the layout of the
partition map.

Ref: NCSDK-30842

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>