scripts: generate psa key attributes strips to many zeros
Turns out that the string strip function in python doesn't just strip
any occurrence of any sub string which is equal to the string I pass
in but also also any occurrence of any of the characters I pass in to
the function. Replaced the strip function with a function that strips
away only the sub-string I pass in to the function from the beginning
of the string.

Signed-off-by: Even Falch-Larsen <even.falch-larsen@nordicsemi.no>