drivers: mpsl: Add flash synchronization via RPC for nRF54 app core
There might be several cases in which it is necessary to synchronize
the MRAM writes requested by the application core with MPSL:

1) A situation where the application core is responsible for writing
to the radio MRAM memory or the application core writes to the part
of the application MRAM memory which resides in the same MRAM bank
as the code currently executed on the radio core

2) For the current use case of the SUIT Flash IPUC driver (application
core requests the Secure Domain to write data to the radio core MRAM)

In these cases, the MRAM writes might stall the radio core execution,
leading to timeouts and soft device controller asserts.

As MPSL is operating on the radio core, the driver introduced by
this commit requests an MPSL timeslot from the radio core via RPC.
The radio core then sends a response with the start of the MPSL
timeslot.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>