nrf_security: Adjust the size of hash operation in Cracen
This updates the hash operation size depending on the enabled
algorithms.

This is an optimization which will reduce stack usage for use
cases which only require a subset of the enabled algorithms.

Notice that the nrf-psa-crypto-user-config.h header is used
directly here instead of the psa/crypto.h. This is done
because including the psa/crypto.h creates circular dependencies
in the headers. The sxsymcrypt is in a lower layer than then PSA
crypto driver layer so it makes sense to avoid going from a lower
layer in the driver to a higher one. The PSA configuration is still
being used since this is the only way to do crypto configuration.
Since the nrf-psa-crypto-user-config.h is a static configuration
file without any dependencies it is reasonable to use it inside
the lower layers of the driver.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>