mpsl: clock_ctrl: For 54H fetch LFCLK accuracy from BICR
This replaces the default LFCLK accuracy passed to MPSL and the
clock driver with what is configured in BICR
instead of using a pessimistic assumption of 500 ppm.

For the 54H DK the board configuration file and corresponding BICR
file enables the use of a LFXO with an accuracy of 20 PPM.
In case the XTAL is not used, the LFRC clock accuracy of 500 ppm will
be used.
See also `zephyr/boards/nordic/nrf54h20dk/bicr.json`.

The code which is added to extract the clock accuracy is obtained
from the `clock_control_nrf2_lfclk.c`. This should eventually be
obtained through a separate API.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>