samples: bluetooth: fast_pair: locator_tag: Use CryptoCell in MCUBoot
Changes the MCUBoot signature type to the Elliptic curve digital
signatures with curve P-256 (ECDSA P256) for nRF52840 devices.
This is done to use Cryptocell 310 for image confirmation. This
change breaks backwards compatibility. It is not possible to
perform DFU from old signature type to new signature type.

Jira: NCSDK-32692

Signed-off-by: Aleksander Strzebonski <aleksander.strzebonski@nordicsemi.no>