suit: Fix reading uninitialized IPUC
Returning valid data on uninitialized IPUC is misleading for the cache
pool implementation.
It may result in a strange cycle, where device assumes that some
payloads are downloded, but once a missing one is fetched, they are
erased.
At this point, manifest logic may assume that all required payloads are
fetched, but in fact the update will be rejected, leading to unnecesary
confusion.
The drawback of returning 0xFF is that after a reboot the device will
always see the cache as empty.

Ref: NCSDK-NONE

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>