applications: nrf5340_audio: Clear bonding only on reset for testing
The CI system cannot reset the bonding on reset with button 5
as the button testing is connected to the scan delegator.

Split the generating a new address and clearing the bonding
on every reset config (TESTING_BLE_ADDRESS_RANDOM) into the
create and bonding clear config (TESTING_BLE_BONDING_CLEAR).

Thus allowing only the bonding information to be cleared.

Signed-off-by: Graham Wacey <graham.wacey@nordicsemi.no>