samples: crypto: add nRF54L15 w/ Oberon to sha256/aes_gcm
Add the nrf54l15dk/nrf54l15/cpuapp board target to the Oberon test
scenarios for the sha256 and aes_gcm samples.
This is to specifically test SW crypto on that board target.

It could be made in a better way and extended to other board
targets/samples but this is a quick addition to get some
SW crypto testing to that board target as part of CI.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>
(cherry picked from commit 57cc0843600fa906b1fe71998c6ffc4d9456b547)
Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>