modules: memfault: Check project key only when used to upload
In cases where an upstream device uploads data to Memfault on behalf
of a downstream device, the project key is not required. This change
allows the project key to be optional in those cases. For the
NCS Memfault sources, this is when we are not using MDS or the
Memfault Zephyr HTTP client.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>
(cherry picked from commit b6f185804b5b65273e7f1d6af1b0d1b873abb1d1)