cracen: moving declaration of variable to top of function
Declaration of variables after a label inside a switch statement is a
c23 extension, not c99.

This results in the following warning when compiling with clang:
> ..subsys/nrf_security/src/drivers/cracen/cracenpsa/src/ecdsa.c:407:3:
>    warning: label followed by a declaration is a C23 extension
>  407 |          size_t copylen = MIN(digestsz, opsz - hmac_op->tlen);
>      |          ^
> 1 warning generated.

There is no practical reason why the variable should be declared
inside the switch statement, therefore move the declaration and place it
together with declaration of other variables.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>