samples: net: https_client: Fix certificate parsing error
Based on [1] the example.com server has updated it's cert to use
secp384r1 curve which isn't enabled in the config, earlier it was not
EC, but RSA, this causes MbedTLS to fail with 0x3A00
(MBEDTLS_ERR_PK_UNKNOWN_NAMED_CURVE).

Fix this by enable the respective curve in both 54L and nRF70, hope this
works automatically with Modem platforms too (nRF91X).

Fixes SHEL-3434.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>