samples: bluetooth: fast_pair: locator_tag: fix DFU for nRF53 NS targets
Fixed the DFU issue for the non-secure targets that are based on the
nRF53 SoC:

- thingy53/nrf5340/cpuapp/ns
- nrf5340dk/nrf5340/cpuapp/ns

Due to the misaligned partition size of the Primary and the Secondary
slots, the MCUboot bootloader was unable to boot the new firmware
correctly.

Aligned the size of the Primary and the Secondary slots to fix this
issue.

Ref: NCSDK-33065

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>
(cherry picked from commit 64c0fc7753dc846e1bf1454c3e06e6bc6aadbf4a)