mpsl: pm: Don't use the same bit for MRAM and PM requests
The LOW_LATENCY_PM_BIT and LOW_LATENCY_MRAM_BIT had the same
value. The MPSL was notified about completion of low latency
request before it actually happened.
The if statement in m_register_latency() for MRAM always returned
true because of use of the same bit number for both constants.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>