samples: bootloader: Align end of S1 partition
As both S0 and S1 partitions are protected using
fprotect_area(), with one call, but not separately,
then we need to ensure that beginning of S0 is aligned
as well as end of S1.

Beginning of S1 partition does not need to be aligned.
This would allow end of S0 and beginning of S1 to be
within a same SPU protection block.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>