tests: settings: performance: fix storage size
When partition manager tries to assign the partition size for the
settings it allocates the minimum size which is 8K bytes.
This is not enough for the performance test.
Replicate the same test from sdk-zephyr and change the partition size
to 32K bytes.

Signed-off-by: Riadh Ghaddab <rghaddab@baylibre.com>
(cherry picked from commit c779cbafb3911df67571143756d3913bc2a5b0c7)