nrf_security: drivers: cracen: cipher: multi-part to return unsupported
Make the multi-part cipher setup functions return
PSA_ERROR_NOT_SUPPORTED on the 54L20 so that the failure reason
is more clear than a wrong ciphertext/cleartext.

The setup function is moved to the top of the file to be callable by all
the other functions as it's now static.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>