nrf_security: Apply is_first_block to Cracen HMAC
The is_first_block variable should be used by the Cracen
HMAC as well as CMAC. So move the variable to the MAC
operation instead of the CMAC specific one.

Also use this variable to HMAC as well. The call sx_hash_resume_state
expects that the sx_hash_save_state was called before so this makes
sure that the sx_hash_resume_state is only called in this case

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>