samples: bluetooth: iso_time_sync: Fix iso_tx.c when asserts enabled
This sample would assert if asserts were enabled, this is
due to calling bt_iso_chan_send/bt_iso_chan_send_ts which
in turn call sched_lock, which is not allowed to be called
from a timer_isr. To fix this, schedule a work queue to run
at the desired time.

Another assert would then occur as we were sending SDUs with
incorrect timestamps due to a race condition between the
iso_sent callback and iso_connected callback meaning we
would add an incorrect timestamp to some sdus. This causes
the controller to return an error and the buffer gets lost,
leading to an assert.

Signed-off-by: Sean Madigan <sean.madigan@nordicsemi.no>