application: serial_lte_modem: sms: Fix receiving concatenated SMS
Collected rsp_buf for received concatenated SMS was from the stack.
When 2nd and following parts were received, the first part was lost.
So we always sent everything except the first part and not
the #XSMS notification name and parameters other than the data.
Moved rsp_buf to static RAM so it's preserved over multiple SMS parts.

Jira: LRCS-169

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>