ext: remove dependency on POSIX_THREADS not being selected
CURL and IPERF3 depend on POSIX_THREADS not being selected but
POSIX_API selects POSIX_BASE_DEFINITIONS and subsequently POSIX_THREADS,
resulting in a dependency conflict.

There is no justification provided by the responsible team as to what
the exact problem was with  PTHREAD_IPC (now POSIX_THREADS) and whatever
it still is or is not.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>