soc: nrf54l: Add the linker script for the KMU push area
The nRF54L devices have a reserved RAM area which is used
by the KMU as a target for the push operation.

This is done by including a linker script which defines this area.
For nRF54L10/L15 this linker script is directly included by the
CMake logic in the soc folder in sdk-zephyr.

Since the nRF54LM20A doesn't use the sdk-zephyr CMake logic
from the soc directory we need to copy the logic here.

When the nRF54LM20A uses sdk-zephyr CMake logic this can be
removed.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>