nrf_security: Fix AEAD issue with ChachaPoly in Cracen
ChachaPoly used sx_hw_reserve instead of sx_aead_hw_reserve.
This was problematic because the sx_aead_hw_reserve is responsible
of calling the prepare_key function pointer. This is relevant to
the KMU keys since they rely on this function to prepare the key
material.

This fixes the issue by using the sx_aead_hw_reserve function
for ChachaPoly the same way that is used by CCM and GCM.

Ref: NCSDK-33315

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>