sysbuild: Migrate from sysbuild_dt_*
The `sysbuild_dt_*` CMake API came from upstream Zephyr PR #: 73903.
This had been in review for almost a year, and in the end, the proposed
API ended up being replaced.

This merits an entry in the migration guide. Note that since the API was
cherry-picked from PR, there's no deprecation process.

Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>