bluetooth: samples: Save and restore local steps in RAS server callback
Because of the way connection callbacks work, and because
bt_le_cs_step_data_parse() consumes the data in the buffer, any user
callbacks called after this module's would end up losing access to the
step data.

To avoid this problem, we can do a simple save and restore when parsing.

Signed-off-by: Olivier Lesage <olivier.lesage@nordicsemi.no>