debug: coredump: fix CRC calcuation for changing data
Fix CRC calculation in the nRF internal flash core dump
backend in the scenario where the data being dumped keeps
changing. This occurs in one of the Core Dump modes that
dumps also the IRQ stack.

The fix is done by writing all data using an intermedate
buffer and doing the CRC calculation on that buffer instead
of input data.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>