net: openthread: rpc: call client with mutex unlocked
OpenThread over RPC server invoked OpenThread callbacks
on the RPC client with the OpenThread stack mutex locked.
This may have lead to a deadlock when the client called the
server while the server was calling the client.

Therefore, unlock the OpenThread stack mutex before calling
the client, and lock it again when the response arrives.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>