dult: Move assert after ready check in ANOS characteristic handler
Moved assert after the subsys ready check in the ANOS GATT
characteristic write handler as the pointer checked in the assert
is initialized on the dult_enable call.
It could result in assertion when the DULT was not yet initialized
but characteristic was accessed as it is statically defined.
Now instead of assertion it just gracefully returns from the function.

Jira: NCSDK-33739

Signed-off-by: Mateusz Kapala <mateusz.kapala@nordicsemi.no>