matter: Refactor Matter samples to start using Snippets
In the new Matter samples structure approach, we rely on the
snippets to make a base for the Matter application configuration.
Snippets that are required by the matter application are added
automatically thanks to the "sysbuild.cmake" file in the sample
directory.

In this approach, we have some new snippets in sdk-nrf:

- matter - which is responsible for setting up the base of the
Matter project and ensuring that all modules required by Matter
are enabled. It also contains specific definitions for all
supported boards.

- matter-low-power - This snippet is used to enable low power mode
in a Matter sample and allow a device to enter sleep mode in idle.

- matter-bootloader - This is a configuration for all supported
boards dedicated to the mcuboot image. This snippet can be used
only with the mcuboot image.

- matter-network-core - This snippet prepares the ipc_radio image
to work with Matter project. This snippet is dedicated only
to devices that have the Network Core.

- matter-release - This snippet prepares a project to release.
Using this snippet, all debug features are disabled, and the
firmware footprint is the lowest.

- matter-internal-dfu - This snippet prepares a Matter project
to use internal memory space to store the DFU image. This snippet
is dedicated only to nRF54L series devices.

The previous functionalities are kept, so to build a project
in the release, you can still add FILE_SUFFIX=release.
The appropriate snippet for release mode will be added
automatically.

This approach allows reducing the number of exactly the same
configuration files, and keeps all the configuration in one
location.

All configs that are set by snippets can be overwritten by
local overlay files, or direct CMAKE build arguments.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>