net: lib: downloader: Don't forward first HTTP data fragment
Do not immediately forward the first fragment of data after
we have parsed the HTTP header. On a worst case, this amount
of data is not enough for FOTA Downloader library to detect
the image type and it would fail.

So when HTTP header is parsed, ask zsock_recv() once more to
fill up the buffer with data if we have less than 32 bytes.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>