drivers: serial: uart_nrf_sw_lpuart: clear req pin at the end of req_pin_idle() function
According to driver doc req pin should be configured as output and set low in idle mode. However inside the lpuart driver pin is only configured as output without clearing. This creates weakness of the driver which can be utlized by simply setting pin high prior to driver initialization. This commit is intended to fix it.

Signed-off-by: Piotr Radecki <piotr.radecki@jrdltd.co.uk>