mpsl: fem: adjust to FEM config API change
In the MPSL the FEM configuration API will change soon.
The change is signalled by presence of `_MPSL_FEM_CONFIG_API_NEXT`
macro.
When it is not defined, the code compiles as it was before.
When it is defined the code is changed as required by the MPSL API:
- For the nRF54L Series the DPPI channels and PPIB channels are now
  allocated internally in MPSL FEMs. The Implementation of
  mpsl_fem_hwres.h API is needed. The following required functions
  are added: `mpsl_fem_hwres_dppi_channel_alloc`,
  `mpsl_fem_hwres_ppib_channel_alloc`
- The MPSL's struct `mpsl_fem_gpiote_pin_config_t` no more contains the
  `ppib_channels` field. The function `mpsl_fem_utils_gpiote_pin_init`
  is no more necessary.

After MPSL containing the changes is present in the  sdk-nrfxlib,
only the code with `defined(_MPSL_FEM_CONFIG_API_NEXT)` will be left.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>