cmake: NCS-only fix for variant configuration with no partition manager
This commit introduces a (probably temporary) way of fixing passing
Kconfigs to a variant image.

In the solution, when not using partition manager, both the .config
and autoconf.h files are copied into the variant image. However,
the CONFIG_FLASH_LOAD_OFFSET and CONFIG_FLASH_LOAD_SIZE are modified
to match the values from zephyr,code-partition - as this are
the values used for linking.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>