samples: crypto: psa_tls: Don't close server socket on client disconnect
Closing the listening server socket when client disconnects isn't really
necessary, moreover adding a 200ms sleep after this creates a window
where no new connection can be accepted, which in turn can be a source
of races during tests.

Therefore, rework the server side of the sample to keep the listening
socket open - it will only be recreated if accepting new connection
fails.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>