applications: serial_lte_modem: Increase CMUX notification buffer
Long AT notifications are truncated with CMUX based on
CONFIG_SLM_CMUX_NOTIFICATION_TX_BUFFER_SIZE which was 256 by default.
This is now increased to match the CONFIG_AT_MONITOR_HEAP_SIZE
which reserves 4kB for the notifications.
We may need to revisit this when doing memory optimizations.

Jira: SLM-98

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>