samples: nrf_rpc: entropy_nrf53: Lower serialization init priority
Lower the serialization initialization priority to the APPLICATION
level. Upstream Zephyr PR #89072 moved part of the nRF53
initializations to the soc_late_init_hook(), which is executed after the
POST_KERNEL initializations. Apparently there was something important
for IPC to work, as the cores failed to communicate if the serialization
was initialized at POST_KERNEL level.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>