drivers: flash_rpc: Lower driver init priority
Some nRF53 initializations are now done after POST_KERNEL init level, so
initialize the serialization at the APPLICATION level.

As DT doesn't allow to initialize devices at the APPLICATION level, make
use of the zephyr,deferred-init property and initialize the device
manually from SYS_INIT().

For whatever reason the property was not recognized if
nordic,rpc-flash-controller binding was no present, so add it.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>