samples: ipc: ipc_service: Add workaround for ICBMSG configurations
Sporadically, IPC Service sample hangs on nRF54LM20A when
ICBMSG is used as backend.
It seems that cpuapp to early tries to establish IPC connection.

Add delay on cpuapp to wait a bit until the remote core
boots and remote is capable to accept IPC connection.

This has negative impact on the firs throughput reported
after reset.

Signed-off-by: Sebastian Głąb <sebastian.glab@nordicsemi.no>