mpsl: fem: adjust to FEM config API change
In the MPSL the FEM configuration API will change soon.
To smoothly transition through the change in public API for FEM config
the new version will define a macro `_MPSL_FEM_CONFIG_API_NEXT`.

The provided code supports both API versions.

When the macro is not defined, the code compiles as it was before.
When the macro is defined the code is changed as required by the new
MPSL API:
- For the nRF54L Series the DPPI channels and PPIB channels for FEMs are
  allocated internally in MPSL. The recently provided implementation
  of functions in mpsl_hwres.h is responsible for that.
- The MPSL's struct `mpsl_fem_gpiote_pin_config_t` no more contains the
  `ppib_channels` field. The function `mpsl_fem_utils_gpiote_pin_init`
  is no more necessary.

When the MPSL containing the changes is present in the sdk-nrfxlib,
only the code variant for `defined(_MPSL_FEM_CONFIG_API_NEXT)` will
be kept.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>