snippets: nordic-bt-rpc: update nrf54h20 memory node layout
Update the memory map overlay to account for changes to the memory
layout coming in with the IronSide support for nrf54h20.
The radio settings partition has been left where it was since it
doesn't conflict with anything, though it probably should be updated
when the new nrf54h20 memory map is a bit more final.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>