samples: smp_svr_mini_boot: Disable Cracen protected RAM config
This disables the configuration: CRACEN_PROVISION_PROT_RAM_INV_DATA
which is used to invalidate the protected RAM after a crypto
operation is finished.
This will make the protected RAM keys unusable for this sample in
MCUBOOT but since they are not used it should not be a problem.

This needs to get disabled because MCUBOOT enables LTO and the
minimal C library which has some compatibility issues at the moment.
Since the protected RAM option uses memset it will result to a failed
build if enabled.

The incompatibility issue with LTO and minimal C is reported here:
https://github.com/zephyrproject-rtos/zephyr/issues/90319

When the Zephyr issue is solved it should be possible to enable
this option if needed later.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>