nrf_cloud_agnss: Increase maximum A-GNSS data size
The maximum A-GNSS data size was too small, if ephemerides
and almanacs were received for all GPS satellites and more than
four QZSS satellites. The planned QZSS constellation size is
seven satellites. The updated size leaves some margin on top
of 32 GPS and 7 QZSS satellites to avoid similar problems in
the future.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>