applications: serial_lte_modem: Fix power pin after idle.
When resuming from idle, we need to wait for the power pin
to resume to nonactive state before re-enabling the power pin
interrupt.

This was already done correctly for power pin toggle, but after
idle, we were reacting to wrong edge, which, in some cases,
caused power off to be called.

Added 50ms timeout to prevent the power button bouncing.
This was observed to sometimes take place when resuming from idle,
which sometimes caused device to power off.

Signed-off-by: Markus Lassila <markus.lassila@nordicsemi.no>