samples: bluetooth: event_trigger: Fix EGU interrupt run kernel sched
Recently MPSL ISRs were fixe to do not trigger kenrel scheduler
because they are zero latency interrupts. The event_trigger sample
uses EGU to be triggered when a connection event happens. The EGU ISR
is connected with IRQ_DIRECT_CONNECT macro but it is defined as
regular ISR (withouth use of macro ISR_DIRECT_DECLARE).

Since there is no return value from the egu_handler() the kenrel
scheduler is not triggered. There is no other event that kicks
the scheduler so sample is hanging in the while loop in the main()
waiting for a NUM_TRIGGERS number of connection events to happen.

Use IRQ_CONNECT macro to connect the egu_handler() as a regular
ISR. That makes sure the kenrel is triggered at end of the ISR.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>