applications: nrf5340_audio: do not force Mbed TLS config file in McuBoot
In the default configuration used in "prj_fota.conf" an RSA key is used
to sign/verify images. "mcuboot-mbedtls-cfg.h" is meant to be used with
EC key types not RSA, so this gives a warning at build time that
CONFIG_MBEDTLS_CFG_FILE was assigned the value 'mcuboot-mbedtls-cfg.h'
but got the value 'config-mbedtls.h.

The simple solution is not to force CONFIG_MBEDTLS_CFG_FILE and let
default values pick the proper configuration file.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>