doc: known_issues: update entry for NCSDK-32268
Updated the known issue for the clock value in the Find My Device
Network (FMDN) module that has incorrect value after the system reboot
in the case of nRF54L devices. The implemented workaround is no longer
needed in the NCS 3.1.0 release and beyond, as the issue with the kernel
function (k_uptime_get) and the kernel uptime has been resolved.

Ref: NCSDK-32514

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>