samples: 802154_phy_test: fix too small system work queue stack size
The PR # 23004 introduced fixed nRF 802.15.4 Radio Driver in which
the field `metadata.data.transmitted.length` is properly set.
Previously this field was set always to zero.
The function `ptt_ctrl_rf_push_packet` was always called with parameter
len=0. With the fix the function is called with proper non-zero length
value. This valid fix revealed a configuration flaw in the
802154_phy_test sample. The system work queue stack size was too small
what caused the application to crash.
This commit fixes this issue.

Signed-off-by: Andrzej Kuros <andrzej.kuros@nordicsemi.no>