tests: pcm_stream_channel_modifier: Interleave and deinterleave support
OCT-3437

As an audio engineer I would like to be able to place a
contiguous buffer representing a channel of audio, into
the specified interleaved channel. Similarly I would
like to take a single channel of interleaved audio and
construct a contiguous buffer of the samples.

The samples will be 8-, 16-, 24- or 32-bit carried
completely in a data item of 8-, 16-, or 32-bits.

Signed-off-by: Graham Wacey <graham.wacey@nordicsemi.no>