dfu_target: Do not erase last page in dfu_target_mcuboot_done
Erasing the last page is already done before writing the trailer
in dfu_target_mcuboot_schedule_update.

It seems the erasing comes from a time a couple of years ago
when the page was not erased inside mcuboot. Currently
the code was not necessary and it caused flash wear out.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>