secure_storage: auto-write HUK when using that as key provider
The HUK library requires the HUK to be written to be able to derive
keys.
When Secure storage uses the HUK library to derive keys, default-enable
the Kconfig option which allows to write the HUK automatically on PSA
Crypto init so that the users don't have to do it themselves and don't
get confused by Secure storage not working out of the box.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>