applications: ipc_radio: Restore previous timeslice size
In the newest Zephyr, there is a new default value for the
CONFIG_TIMESLICE_SIZE Kconfig. Previously, it was set to 0, which
means "infinitely large time slice", but it was changed to the
20 ms according to the approximation of the historical clamping
behaviour.

When both Bluetooth LE and 802.15.4 work, this 20 ms causes
random crashes on the nRF5340 network core. To avoid such
behaviours, we can enforce the previous value and set the
"infinitely large time slice".

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>
(cherry picked from commit 15d1605416f56a5053397351ccf934d623c63b05)