actions: manifest: Remove unnecessary west update command
In commit f7eb8c8, we introduced the use of the use-tree-checkout
option. But at the same time we also restricted the parsed imports to
self:, which means that the import statement for sdk-zephyr in the main
west.yml file in sdk-nrf is ignored when running this action. Hence,
there's no need to actually clone the zephyr tree as part of the action
execution.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>