applications: nrf_desktop: fix memory map for nRF54L10
Fixed the memory layout for the nrf54l15dk/nrf54l10/cpuapp target
in the nRF Desktop application. Previously, it was assumed that
the memory size was 10 KB larger than the actual one. The NVM size
in the nRF54L10 SoC is equal to 1012 KB.

This change in the nRF54L10 partition map is a breaking change and
cannot be performed using DFU. As a result, the DFU procedure will
fail if you attempt to upgrade the sample firmware based on one of
the nRF Connect SDK v3.0 releases.

Ref: NCSDK-34584

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>