treewide: use proper placeholder for invalid sec_tag
The fact that the modem uses uint32_t for sec_tags,
while Zephyr uses int, is a bit unfortunate.
Often, -1 or 0 are used to indicate an invalid sec_tag,
and it is checked whether a sec_tag is positive.

However, there are some "debug" sec_tags for the modem,
that register as negative values, while being valid.
To avoid confusion, use a proper placeholder for an invalid sec_tag,
which is NRF_SEC_TAG_TLS_INVALID (0xFFFFFFFF) for the modem,
and 0xFFFFFFFF for libraries not specific to nRF91 modems.

Signed-off-by: Maximilian Deubel <maximilian.deubel@nordicsemi.no>